% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{add_sample_annotations}
\alias{add_sample_annotations}
\title{Adds sample annotation to musica object with available samples}
\usage{
add_sample_annotations(
  musica,
  annotations,
  sample_column = "Sample_ID",
  columns_to_add = colnames(annotations)
)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{annotations}{table of sample-level annotations to add}

\item{sample_column}{name of sample name column}

\item{columns_to_add}{which annotation columns to add, defaults to all}
}
\value{
Sets sample_annotations slot {no return}
}
\description{
Adds sample annotation to musica object with available samples
}
\examples{
musica <- readRDS(system.file("testdata", "musica_sbs96.rds", package = "musicatk"))
init_sample_annotations(musica)
sample_annotations <- read.table(system.file("testdata",
"sample_annotations.txt", package = "musicatk"), sep = "\t", header=TRUE)
add_sample_annotations(musica = musica, annotations = sample_annotations,
sample_column = "Sample_Names", columns_to_add = "Tumor_Subtypes")
musica
}
