% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{add_variant_type}
\alias{add_variant_type}
\title{Generates a variant type table}
\usage{
add_variant_type(tab)
}
\arguments{
\item{tab}{Input variant table}
}
\value{
Returns the inputted variant table with variant type ("SBS", "DBS",
"INS", "DEL") added as an appended "Variant_Type" column
}
\description{
Generates a variant type table
}
\examples{
data(musica)
variants <- variants(musica)
musicatk:::add_variant_type(variants)
}
\keyword{internal}
