% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{adjustment_threshold}
\alias{adjustment_threshold}
\alias{adjustment_threshold,single_benchmark-method}
\alias{adjustment_threshold<-}
\alias{adjustment_threshold<-,single_benchmark,numeric-method}
\title{Retrieve the adjustment_threshold from a single_benchmark object}
\usage{
adjustment_threshold(x)

\S4method{adjustment_threshold}{single_benchmark}(x)

adjustment_threshold(x) <- value

\S4method{adjustment_threshold}{single_benchmark,numeric}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark}} object.}

\item{value}{The adjustment_threshold}
}
\value{
The adjustment_threshold value.
}
\description{
The \code{adjustment_threshold} is a cosine similarity value of
high confidence.Comparisons that meet this cutoff are assumed to be likely,
while those that fall below the cutoff will be disregarded if the predicted
signature is already captured above the threshold.
}
