% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_transcript_strand}
\alias{annotate_transcript_strand}
\title{Add transcript strand annotation to SBS variants (defined in genes only)}
\usage{
annotate_transcript_strand(musica, genome_build, build_table = TRUE)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{genome_build}{Which genome build to use: hg19, hg38, or a custom TxDb
object}

\item{build_table}{Automatically build a table from this annotation}
}
\value{
None
}
\description{
Add transcript strand annotation to SBS variants (defined in genes only)
}
\examples{
data(musica)
annotate_transcript_strand(musica, 19)
}
