% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variant_length}
\alias{annotate_variant_length}
\title{Adds an annotation to the input musica's variant table with length of each
variant}
\usage{
annotate_variant_length(musica)
}
\arguments{
\item{musica}{Input samples}
}
\value{
None
}
\description{
Adds an annotation to the input musica's variant table with length of each
variant
}
\examples{
data(musica)
annotate_variant_length(musica)
musica
}
