% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variant_type}
\alias{annotate_variant_type}
\title{Annotate variants with variant type ("SBS", "INS", "DEl", "DBS")}
\usage{
annotate_variant_type(musica)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}
}
\value{
None
}
\description{
Annotate variants with variant type ("SBS", "INS", "DEl", "DBS")
}
\examples{
musica <- readRDS(system.file("testdata", "musica.rds", package = "musicatk"))
annotate_variant_type(musica)
}
