% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{benchmark_get_comparison}
\alias{benchmark_get_comparison}
\title{Get benchmark comparison table}
\usage{
benchmark_get_comparison(full_benchmark, method_id, prediction)
}
\arguments{
\item{full_benchmark}{The \code{\linkS4class{full_benchmark}} object for the
benchmarking analysis}

\item{method_id}{The identifier for the \code{\linkS4class{single_benchmark}}
object containing the comparison of interest}

\item{prediction}{\code{"Initial"} for the comparison before any benchmarking
adjustments have been made, \code{"Intermediate"} for the comparison after
duplicates have been adjusted but before composites are adjusted, or
\code{"Final"} for the comparison at the end of the benchmarking adjustments.}
}
\value{
A data.frame containing the comparison between true and predicted
signatures
}
\description{
After a prediction has been benchmarked with the \link{benchmark} function,
this function can be used to extract the comparison table between true and
predicted signatures from any step in the benchmarking process.
}
\examples{
data(full_benchmark_example)
final_comparison <- benchmark_get_comparison(full_benchmark_example, "example_k8", "Final")

}
