% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{benchmark_get_entry}
\alias{benchmark_get_entry}
\title{Get a single_benchmark object}
\usage{
benchmark_get_entry(full_benchmark, method_id)
}
\arguments{
\item{full_benchmark}{The \code{\linkS4class{full_benchmark}} object that contains
the desired \code{\linkS4class{single_benchmark}} object}

\item{method_id}{The identifier for the desired \code{\linkS4class{single_benchmark}}
object}
}
\value{
A \code{\linkS4class{single_benchmark}} object
}
\description{
Access a \code{\linkS4class{single_benchmark}} object containing information from
an individual benchmark run from a \code{\linkS4class{full_benchmark}} object.
}
\examples{
data(full_benchmark_example)
indv_benchmark <- benchmark_get_entry(full_benchmark_example, "example_k8")

}
