% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{benchmark_get_prediction}
\alias{benchmark_get_prediction}
\title{Get a benchmark prediction}
\usage{
benchmark_get_prediction(indv_benchmark, prediction)
}
\arguments{
\item{indv_benchmark}{A \code{\linkS4class{single_benchmark}} object containing the
desired prediction. This can be accessed using the \link{benchmark_get_entry}
function.}

\item{prediction}{\code{"Initial"} for the prediction before any benchmarking
adjustments have been made, \code{"Intermediate"} for the prediction after
duplicates have been adjusted but before composites are adjusted, or
\code{"Final"} for the prediction at the end of the benchmarking adjustments.}
}
\value{
A \code{\linkS4class{result_model}} object
}
\description{
Access a \code{\linkS4class{result_model}} object containing a particular prediction
from a \code{\linkS4class{single_benchmark}} object.
}
\examples{
data(full_benchmark_example)
indv_benchmark <- benchmark_get_entry(full_benchmark_example, "example_k8")
initial_prediction <- benchmark_get_prediction(indv_benchmark, "Initial")

}
