% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{benchmark_plot_exposures}
\alias{benchmark_plot_exposures}
\title{Plot exposure comparison from a benchmarking analysis}
\usage{
benchmark_plot_exposures(full_benchmark, method_id, prediction)
}
\arguments{
\item{full_benchmark}{The \code{\linkS4class{full_benchmark}} object for the
benchmarking analysis}

\item{method_id}{The identifier for the \code{\linkS4class{single_benchmark}}
object of interest}

\item{prediction}{\code{"Initial"} for the exposures before any benchmarking
adjustments have been made, \code{"Intermediate"} for the exposures after
duplicates have been adjusted but before composites are adjusted, or
\code{"Final"} for the exposures at the end of the benchmarking adjustments.}
}
\value{
Generates a  ggplot object
}
\description{
After a prediction has been benchmarked with the \link{benchmark} function,
the comparison between the true and predicted exposures at any stage of the
benchmarking process can be plotted.
}
\examples{
data(full_benchmark_example)
benchmark_plot_exposures(full_benchmark_example, "example_k8", "Final")

}
