% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_musica.R
\name{built_tables}
\alias{built_tables}
\alias{built_tables,musica-method}
\title{Retrieve the names of count_tables from a musica object}
\usage{
built_tables(object)

\S4method{built_tables}{musica}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica_from_variants} or \link{create_musica_from_counts}
function.}
}
\value{
The names of created count_tables
}
\description{
The \code{count_tables} contains standard and/or custom
count tables created from variants
}
\examples{
data(res)
built_tables(res)
}
