% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{cluster_exposure}
\alias{cluster_exposure}
\title{Perform k-means clustering from a musica result object}
\usage{
cluster_exposure(
  result,
  nclust,
  method = "kmeans",
  iter.max = 10,
  clara.samples = 5
)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{nclust}{Pre-defined number of clusters.}

\item{method}{Clustering algorithms. Options are "kmeans" (K-means), "hkmeans" (hybrid of hierarchical 
K-means), "hclust" (hierarchical clustering), "pam" (PAM), and "clara" (Clara).}

\item{iter.max}{Maximum number of iterations. Default is 10.}

\item{clara.samples}{Number of samples to be drawn from dataset. Only used when "clara" is selected. 
Default is 5.}
}
\value{
A one-column data frame with sample IDs as row names and cluster number for each sample.
}
\description{
Proportional sample exposures will be used as input to perform clustering.
}
\examples{
set.seed(123)
data(res_annot)
clust_out <- cluster_exposure(res_annot, nclust = 2)
}
\seealso{
\link[stats]{kmeans}
}
