% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{cluster_kmeans}
\alias{cluster_kmeans}
\title{Perform k-means clustering from a musica result object}
\usage{
cluster_kmeans(result, ...)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{...}{Arguments of \code{\link[stats]{kmeans}} function.}
}
\value{
A one-column data frame with sample IDs as row names and cluster number for each sample.
}
\description{
Proportional sample exposures will be used as input for the 
\code{\link[stats]{kmeans}} function to perform clustering.
}
\examples{
set.seed(123)
data(res_annot)
k_out <- cluster_kmeans(res_annot, centers = 2)
}
\seealso{
\link[stats]{kmeans}
}
