% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_results.R
\name{compare_cosmic_v3}
\alias{compare_cosmic_v3}
\title{Compare a result object to COSMIC V3 Signatures; Select exome or genome for
SBS and only genome for DBS or Indel classes}
\usage{
compare_cosmic_v3(
  result,
  variant_class,
  sample_type,
  metric = "cosine",
  threshold = 0.9,
  result_name = deparse(substitute(result)),
  show_plot = TRUE
)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object.}

\item{variant_class}{Compare to SBS, DBS, or Indel}

\item{sample_type}{exome (SBS only) or genome}

\item{metric}{One of \code{"cosine"} for cosine similarity or \code{"jsd"} 
for 1 minus the Jensen-Shannon Divergence. Default \code{"cosine"}.}

\item{threshold}{threshold for similarity}

\item{result_name}{title for plot user result signatures}

\item{show_plot}{Toggle display of plot, or only output table}
}
\value{
Returns the comparisons
}
\description{
Compare a result object to COSMIC V3 Signatures; Select exome or genome for
SBS and only genome for DBS or Indel classes
}
\examples{
data(res)
compare_cosmic_v3(res, "SBS", "genome", threshold = 0.8)
}
