% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic_data.R
\docType{data}
\name{cosmic_v2_sigs}
\alias{cosmic_v2_sigs}
\title{COSMIC v2 SBS96 Signatures Result Object}
\format{
An object of class \code{\linkS4class{"musica_result"}}
See [predict_exposure()].
}
\source{
COSMIC v2, <https://cancer.sanger.ac.uk/cosmic/signatures_v2>
}
\usage{
data(cosmic_v2_sigs)
}
\description{
Data from COSMIC formatted to be used for prediction with individual tumors
and cohorts.
}
\references{
Alexandrov, L., Nik-Zainal, S., Wedge, D. et al. (2013)
Signatures of mutational processes in human cancer. Nature 500, 415–421
([Nature](https://www.ncbi.nlm.nih.gov/pubmed/23945592))
}
\keyword{datasets}
