% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_prediction.R
\name{cosmic_v2_subtype_map}
\alias{cosmic_v2_subtype_map}
\title{Input a cancer subtype to return a list of related COSMIC signatures}
\usage{
cosmic_v2_subtype_map(tumor_type)
}
\arguments{
\item{tumor_type}{Cancer subtype to view related signatures}
}
\value{
Returns signatures related to a partial string match
}
\description{
Input a cancer subtype to return a list of related COSMIC signatures
}
\examples{
cosmic_v2_subtype_map ("lung")
}
