% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic_data.R
\docType{data}
\name{cosmic_v3_sbs_sigs_exome}
\alias{cosmic_v3_sbs_sigs_exome}
\title{COSMIC v3 SBS96 Exome Signatures Result Object}
\format{
An object of class \code{musica_result}.
See [predict_exposure()].
}
\source{
COSMIC v3, <https://cancer.sanger.ac.uk/cosmic/signatures>
}
\usage{
data(cosmic_v3_sbs_sigs_exome)
}
\description{
Data from COSMIC formatted to be used for prediction with individual tumors
and cohorts.
}
\references{
Alexandrov, L.B., Kim, J., Haradhvala, N.J. et al. (2020)
The repertoire of mutational signatures in human cancer. Nature 578, 94–101
([Nature](https://doi.org/10.1038/s41586-020-1943-3))
}
\keyword{datasets}
