% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{create_benchmark}
\alias{create_benchmark}
\title{Create a full_benchmark object}
\usage{
create_benchmark(true_signatures, true_exposures, count_table, variant_class)
}
\arguments{
\item{true_signatures}{A matrix of true signatures by mutational motifs}

\item{true_exposures}{A matrix of samples by true signature weights}

\item{count_table}{Summary table with per-sample un-normalized motif counts}

\item{variant_class}{Mutations are SBS, DBS, or Indel.}
}
\value{
A \code{\linkS4class{full_benchmark}} object
}
\description{
Initialize a \code{\linkS4class{full_benchmark}} object for benchmarking
}
\examples{
data(synthetic_breast_counts)
data(synthetic_breast_true_exposures)
true_sigs <- c("SBS1", "SBS2", "SBS3", "SBS8", "SBS13", "SBS17", "SBS18", "SBS26")
true_signatures <- signatures(cosmic_v2_sigs)[,true_sigs]
full_benchmark <- create_benchmark(true_signatures, 
  synthetic_breast_true_exposures, synthetic_breast_counts, "SBS96")

}
