% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_dbs78_color_schema}
\alias{create_dbs78_color_schema}
\title{Creates color schema for standard doublet base subsitution (DBS)}
\usage{
create_dbs78_color_schema()
}
\value{
Returns the created DBS83 color schema object
}
\description{
Creates color schema for standard doublet base subsitution (DBS)
}
