% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_dbs_table}
\alias{create_dbs_table}
\title{Creates and adds a table for standard doublet base subsitution (DBS)}
\usage{
create_dbs_table(musica, overwrite = overwrite)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica_data}} object.}

\item{overwrite}{Overwrite existing count table}
}
\value{
Returns the created DBS table object
}
\description{
Creates and adds a table for standard doublet base subsitution (DBS)
}
