% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{create_musica_result}
\alias{create_musica_result}
\title{Create a musica_result object}
\usage{
create_musica_result(signatures, exposures, count_table, algorithm = NULL)
}
\arguments{
\item{signatures}{A matrix or data.frame of signatures by mutational motifs}

\item{exposures}{A matrix or data.frame of samples by signature weights}

\item{count_table}{Summary table with per-sample unnormalized motif counts}

\item{algorithm}{Describes how the signatures/weights were generated}
}
\value{
A \linkS4class{musica_result} object
}
\description{
This function creates a \linkS4class{musica_result} object from signatures,
exposures, and a mutation count table.
}
