% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{create_result_model}
\alias{create_result_model}
\title{Load an external model into a result_model object}
\usage{
create_result_model(signatures, exposures, model_id, modality)
}
\arguments{
\item{signatures}{A matrix or data.frame of signatures by mutational motifs}

\item{exposures}{A matrix or data.frame of samples by signature weights}

\item{model_id}{Name of model}

\item{modality}{Modality of the model}
}
\value{
A \linkS4class{result_model} object
}
\description{
This function creates a \linkS4class{result_model} object from signatures,
exposures, and a mutation count table.
}
\examples{
signatures <- signatures(res, "result", "SBS96", "res")
exposures <- exposures(res, "result", "SBS96", "res")
model <- create_result_model(signatures, exposures, "example_model", "SBS96")
}
