% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_sbs192_color_schema}
\alias{create_sbs192_color_schema}
\title{Generate standard SBS192 color schema
using transcript strand}
\usage{
create_sbs192_color_schema(strand_type)
}
\arguments{
\item{strand_type}{Transcript_Strand or Replication_Strand}
}
\value{
Returns the created SBS192 color schema object built using either
transcript strand or replication strand
}
\description{
Generate standard SBS192 color schema
using transcript strand
}
