% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{create_sbs96_color_schema}
\alias{create_sbs96_color_schema}
\title{Generate standard SBS96 color schema}
\usage{
create_sbs96_color_schema()
}
\value{
Returns the created SBS96 color schema object
}
\description{
Generate standard SBS96 color schema
}
