% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_umap}
\alias{create_umap}
\title{Create a UMAP data.frame from a result object}
\usage{
create_umap(
  result,
  annotation,
  n_neighbors = 30,
  min_dist = 0.75,
  spread = 1,
  proportional = TRUE,
  seed = FALSE
)
}
\arguments{
\item{result}{S4 Result Object}

\item{annotation}{Annotation to use for umap coloring}

\item{n_neighbors}{The size of local neighborhood used for views of
manifold approximation. Larger values result in more global the manifold,
while smaller values result in more local data being preserved. Default 30.
See `?uwot::umap` for more information.}

\item{min_dist}{The effective minimum distance between embedded points.
Smaller values will result in a more clustered/clumped embedding where
nearby points on the manifold are drawn closer together, while larger
values will result on a more even dispersal of points. Default 0.2.
See `?uwot::umap` for more information.}

\item{spread}{The effective scale of embedded points. In combination with
‘min_dist’, this determines how clustered/clumped the embedded points are.
Default 1.
See `?uwot::umap` for more information.}

\item{proportional}{Whether weights are normalized to sum to 1 or not}

\item{seed}{Use a seed for reproducible results}
}
\value{
UMAP data.frame
}
\description{
Create a UMAP data.frame from a result object
}
\examples{
result <- readRDS(system.file("testdata", "res_annot.rds",
package = "musicatk"))
create_umap(result = result, annotation = "Tumor_Subtypes",
n_neighbors = 5)
}
