% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{dbs_musica}
\alias{dbs_musica}
\title{dbs_musica}
\format{
An object of class \code{musica}
See [create_musica_from_variants()] and [create_musica_from_counts()].
}
\usage{
data(dbs_musica)
}
\description{
A musica created for testing that includes DBS variants
}
\keyword{datasets}
