% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{description}
\alias{description}
\alias{description,single_benchmark-method}
\alias{description<-}
\alias{description<-,single_benchmark,CharOrNULL-method}
\title{Retrieve the description from a single_benchmark object}
\usage{
description(x)

\S4method{description}{single_benchmark}(x)

description(x) <- value

\S4method{description}{single_benchmark,CharOrNULL}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark}} object.}

\item{value}{The description}
}
\value{
The description
}
\description{
The \code{description} contains further details about the
prediction being benchmarked.
}
