% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{.jsd}
\alias{.jsd}
\title{Calculates 1 - Jensen-Shannon Divergences between all pairs of columns 
between two matrices}
\usage{
.jsd(p, q, epsilon = 1e-07)
}
\arguments{
\item{p}{First matrix}

\item{q}{Second matrix}
}
\value{
Returns matrix of 1 - Jensen-Shannon Divergences
}
\description{
Calculates 1 - Jensen-Shannon Divergences between all pairs of columns 
between two matrices
}
\examples{
p <- c(0.2, 0.3, 0.4, 0.5, 0.6)
q <- c(0.6, 0.5, 0.4, 0.3, 0.2)
.jsd(p, q)
}
