% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{drop_annotation}
\alias{drop_annotation}
\title{Drops a column from the variant table that the user no longer needs}
\usage{
drop_annotation(musica, column_name)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica_data}} object.}

\item{column_name}{Name of column to drop}
}
\value{
None
}
\description{
Drops a column from the variant table that the user no longer needs
}
\examples{
musica <- readRDS(system.file("testdata", "musica.rds", package = "musicatk"))
drop_annotation(musica, "Variant_Type")
}
