% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{example_predicted_sigs}
\alias{example_predicted_sigs}
\title{example_predicted_sigs}
\format{
An object of class \code{matrix}
}
\usage{
data(example_predicted_sigs)
}
\description{
A matrix containing the predicted signatures for a synthetic breast cancer
dataset with 214 samples, generated using NMF and k=8 signatures.
}
\keyword{datasets}
