% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{extract_count_tables}
\alias{extract_count_tables}
\title{Extract count tables list from a musica object}
\usage{
extract_count_tables(musica)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}
}
\value{
List of count tables objects
}
\description{
Extract count tables list from a musica object
}
