% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{final_comparison}
\alias{final_comparison}
\alias{final_comparison,single_benchmark-method}
\alias{final_comparison<-}
\alias{final_comparison<-,single_benchmark,data.frame-method}
\title{Retrieve final comparison between the prediction and the ground
truth from a single_benchmark object}
\usage{
final_comparison(x)

\S4method{final_comparison}{single_benchmark}(x)

final_comparison(x) <- value

\S4method{final_comparison}{single_benchmark,data.frame}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark}} object.}

\item{value}{A data.frame containing the comparisons between the intial
prediction and the ground truth.}
}
\value{
A data.frame containing the comparisons.
}
\description{
The \code{final_comparison} contains the comparison between
the ground truth and the final prediction, after duplicates and composites
are corrected.
}
