% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{final_pred}
\alias{final_pred}
\alias{final_pred,single_benchmark-method}
\alias{final_pred<-}
\alias{final_pred<-,single_benchmark,result_model-method}
\title{Retrieve final prediction from a single_benchmark object}
\usage{
final_pred(x)

\S4method{final_pred}{single_benchmark}(x)

final_pred(x) <- value

\S4method{final_pred}{single_benchmark,result_model}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark}} object.}

\item{value}{A \code{\linkS4class{result_model}} object containing the
signatures and exposures from the final prediction.}
}
\value{
An object of class \code{\linkS4class{result_model}} containing
signatures and exposures from the final prediction.
}
\description{
The \code{final_prediction} contains the signatures and
exposures after duplicates and composites are corrected.
}
