% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_full_benchmark.R
\docType{class}
\name{full_benchmark-class}
\alias{full_benchmark-class}
\title{Object that contains information for a full benchmarking analysis where
multiple predictions may be benchmarked}
\description{
Object that contains information for a full benchmarking analysis where
multiple predictions may be benchmarked
}
\section{Slots}{

\describe{
\item{\code{ground_truth}}{A \code{\linkS4class{musica}} object containing the true
signatures and exposures.}

\item{\code{method_view_summary}}{A matrix containing summary information across all
benchmark runs, organized by run.}

\item{\code{sig_view_summary}}{A matrix containing summary information across all
benchmark runs, organized by signature.}

\item{\code{indv_benchmarks}}{A list of \code{\linkS4class{single_benchmark}} objects,
each containing information from a single benchmark run.}
}}

