% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{full_benchmark_example}
\alias{full_benchmark_example}
\title{full_benchmark_example}
\format{
An object of class \code{full_benchmark}
}
\usage{
data(full_benchmark_example)
}
\description{
A \code{full_benchmark} object containing ground truth signatures and exposures
from a synthetic breast cancer dataset with 8 known signatures as well as
a disocvery prediction that has been benchmarked.
}
\keyword{datasets}
