% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_count_table.R
\name{get_count_table}
\alias{get_count_table}
\title{Retrieve count_table matrix from count_table object}
\usage{
get_count_table(count_table)
}
\arguments{
\item{count_table}{A \code{\linkS4class{count_table}} object.}
}
\value{
The count table
}
\description{
The count table
}
