% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_result_collection.R
\name{get_modality}
\alias{get_modality}
\alias{get_modality,musica-method}
\alias{get_modality,result_collection-method}
\title{Retrieve modality from a result collection object}
\usage{
get_modality(x, ...)

\S4method{get_modality}{musica}(x, result, modality)

\S4method{get_modality}{result_collection}(x, modality)
}
\arguments{
\item{x}{The the object generated by this package}
}
\value{
A list of modality which contains models results object
}
\description{
\code{modality} a list contains model results for different modality
}
