% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_musica}
\alias{get_musica}
\alias{get_musica,musica_result-method}
\title{Retrieve musica from a musica_result object}
\usage{
get_musica(result)

\S4method{get_musica}{musica_result}(result)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}
}
\value{
A \code{\linkS4class{musica}} musica object
}
\description{
The \code{\linkS4class{musica}} musica contains variants, 
count tables, and sample annotations
}
\examples{
data(res)
get_musica(res)
}
