% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_musica.R
\name{get_result_list_entry}
\alias{get_result_list_entry}
\alias{get_result_list_entry,musica,character-method}
\title{Retrieve result_list entry from a musica object}
\usage{
get_result_list_entry(object, result_name)

\S4method{get_result_list_entry}{musica,character}(object, result_name)
}
\arguments{
\item{object}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica_from_variants} or \link{create_musica_from_counts}
function.}

\item{result_name}{The name of the result_list entry.}
}
\value{
A list of results
}
\description{
The \code{result_list} contains results from various runs
}
\examples{
data(res)
get_result_list_entry(res, "result")
}
