% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{get_sample_annotations}
\alias{get_sample_annotations}
\title{Return sample annotation from musica object}
\usage{
get_sample_annotations(musica)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}
}
\value{
Sets sample_annotations slot {no return}
}
\description{
Return sample annotation from musica object
}
\examples{
musica <- readRDS(system.file("testdata", "musica.rds", package = "musicatk"))
init_sample_annotations(musica)
get_sample_annotations(musica)
}
