% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{get_sample_names}
\alias{get_sample_names}
\title{Return samples names for musica object}
\usage{
get_sample_names(musica)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica_data}} object.}
}
\value{
Returns names of samples in musica object
}
\description{
Return samples names for musica object
}
\examples{
musica <- readRDS(system.file("testdata", "musica.rds", package = "musicatk"))
get_sample_names(musica)
}
