% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_full_benchmark.R
\name{ground_truth}
\alias{ground_truth}
\alias{ground_truth,full_benchmark-method}
\alias{ground_truth<-}
\alias{ground_truth<-,full_benchmark,musica-method}
\title{Retrieve the ground_truth from a full_benchmark object}
\usage{
ground_truth(x)

\S4method{ground_truth}{full_benchmark}(x)

ground_truth(x) <- value

\S4method{ground_truth}{full_benchmark,musica}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{full_benchmark}} object.}

\item{value}{A \code{\linkS4class{musica}} object containing the ground truth.}
}
\value{
A \code{\linkS4class{musica}} object containing the ground truth.
}
\description{
The \code{ground_truth} is a \code{\linkS4class{musica}} object
that contains the true signatures and exposures.
}
