% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_full_benchmark.R
\name{indv_benchmarks}
\alias{indv_benchmarks}
\alias{indv_benchmarks,full_benchmark-method}
\alias{indv_benchmarks<-}
\alias{indv_benchmarks<-,full_benchmark,list-method}
\title{Retrieve the indv_benchmarks list from a full_benchmark object}
\usage{
indv_benchmarks(x)

\S4method{indv_benchmarks}{full_benchmark}(x)

indv_benchmarks(x) <- value

\S4method{indv_benchmarks}{full_benchmark,list}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{full_benchmark}} object.}

\item{value}{A list of \code{\linkS4class{single_benchmark}} objects.}
}
\value{
A list of \code{\linkS4class{single_benchmark}} objects
}
\description{
The \code{indv_benchmarks} list is a list of
\code{\linkS4class{single_benchmark}} objects, each containing information
about a single benchmark run.
}
