% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{init_sample_annotations}
\alias{init_sample_annotations}
\title{Initialize sample annotation data.table with sample names from variants}
\usage{
init_sample_annotations(musica)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}
}
\value{
Sets sample_annotations slot {no return}
}
\description{
Initialize sample annotation data.table with sample names from variants
}
\examples{
musica <- readRDS(system.file("testdata", "musica_sbs96.rds", package = "musicatk"))
init_sample_annotations(musica)
}
