% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{initial_comparison}
\alias{initial_comparison}
\alias{initial_comparison,single_benchmark-method}
\alias{initial_comparison<-}
\alias{initial_comparison<-,single_benchmark,data.frame-method}
\title{Retrieve initial comparison between the prediction and the ground
truth from a single_benchmark object}
\usage{
initial_comparison(x)

\S4method{initial_comparison}{single_benchmark}(x)

initial_comparison(x) <- value

\S4method{initial_comparison}{single_benchmark,data.frame}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark}} object.}

\item{value}{A data.frame containing the comparisons between the intial
prediction and the ground truth.}
}
\value{
A data.frame containing the comparisons.
}
\description{
The \code{initial_comparison} contains the comparison between
the ground truth and the initial prediction, before duplicates or composites
are corrected.
}
