% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{initial_pred}
\alias{initial_pred}
\alias{initial_pred,single_benchmark-method}
\alias{initial_pred<-}
\alias{initial_pred<-,single_benchmark,result_model-method}
\title{Retrieve initial prediction from a single_benchmark object}
\usage{
initial_pred(x)

\S4method{initial_pred}{single_benchmark}(x)

initial_pred(x) <- value

\S4method{initial_pred}{single_benchmark,result_model}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark} }object.}

\item{value}{A \code{\linkS4class{result_model}} object containing the
signatures and exposures from the initial prediction.}
}
\value{
An object of class \code{\linkS4class{result_model}} containing
signatures and exposures from the initial prediction, before duplicates
and composites are corrected.
}
\description{
The \code{initial_prediction} contains the signatures and
exposures before duplicates or composites are corrected.
}
