% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{intermediate_comparison}
\alias{intermediate_comparison}
\alias{intermediate_comparison,single_benchmark-method}
\alias{intermediate_comparison<-}
\alias{intermediate_comparison<-,single_benchmark,data.frame-method}
\title{Retrieve intermediate comparison between the prediction and the ground
truth from a single_benchmark object}
\usage{
intermediate_comparison(x)

\S4method{intermediate_comparison}{single_benchmark}(x)

intermediate_comparison(x) <- value

\S4method{intermediate_comparison}{single_benchmark,data.frame}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark}} object.}

\item{value}{A data.frame containing the comparisons between the intermediate
prediction and the ground truth.}
}
\value{
A dataframe containing the comparisons.
}
\description{
The \code{intermediate_comparison} contains the comparison
between the ground truth and the intermediate prediction, after duplicates
are corrected but before composites are corrected.
}
