% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{intermediate_pred}
\alias{intermediate_pred}
\alias{intermediate_pred,single_benchmark-method}
\alias{intermediate_pred<-}
\alias{intermediate_pred<-,single_benchmark,result_model-method}
\title{Retrieve intermediate prediction from a single_benchmark object}
\usage{
intermediate_pred(x)

\S4method{intermediate_pred}{single_benchmark}(x)

intermediate_pred(x) <- value

\S4method{intermediate_pred}{single_benchmark,result_model}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark}} object.}

\item{value}{A \code{\linkS4class{result_model}} object containing the
signatures and exposures from the intermediate prediction.}
}
\value{
An object of class \code{\linkS4class{result_model}} containing
signatures and exposures from the intermediate prediction.
}
\description{
The \code{intermediate_prediction} contains the signatures and
exposures after duplicates are corrected but before composites are corrected.
}
