% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{jsd}
\alias{jsd}
\title{Compare two vectors similarity based on Jensen-Shannon Divergence}
\usage{
jsd(p, q)
}
\arguments{
\item{p}{First vector}

\item{q}{Second vector}
}
\value{
Returns Jensen-Shannon Divergence of the two vectors
}
\description{
Compare two vectors similarity based on Jensen-Shannon Divergence
}
\examples{
p <- c(0.2, 0.3, 0.4, 0.5, 0.6)
q <- c(0.6, 0.5, 0.4, 0.3, 0.2)
jsd(p, q)
}
