% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{k_select}
\alias{k_select}
\title{Plots for helping decide number of clusters for k-means}
\usage{
k_select(result, method = c("elbow", "silhouette", "gap"), n = 10)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{method}{A single character string indicating which statistic to use for plot. 
Options are "elbow" (total within cluster sum of squares), "silhouette" (average silhouette 
coefficient), and "gap" (gap statistic). Default is "elbow".}

\item{n}{An integer indicating maximum number of clusters to test. Default is 10.}
}
\value{
A ggplot object.
}
\description{
To help decide the number of cluster for k-means, three different methods 
are provided: total within cluster sum of squares, average silhouette coefficient, 
and gap statistics.
}
\examples{
data(res_annot)
#Make an elbow plot
withr::with_seed(123, k_select(res_annot, method = "elbow", n = 6))
#Plot average silhouette coefficient against number of clusters
withr::with_seed(123, k_select(res_annot, method = "silhouette", n = 6))
#Plot gap statistics against number of clusters
withr::with_seed(123, k_select(res_annot, method = "gap", n =6))
}
\seealso{
\link[factoextra]{fviz_nbclust}
}
