% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{method_id}
\alias{method_id}
\alias{method_id,single_benchmark-method}
\alias{method_id<-}
\alias{method_id<-,single_benchmark,character-method}
\title{Retrieve the method id from a single_benchmark object}
\usage{
method_id(x)

\S4method{method_id}{single_benchmark}(x)

method_id(x) <- value

\S4method{method_id}{single_benchmark,character}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark}} object.}

\item{value}{An identifier for the benchmark run.}
}
\value{
A matrix containing the summary.
}
\description{
The \code{method_id} is the identifier for the benchmark run.
}
