% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_full_benchmark.R
\name{method_view_summary}
\alias{method_view_summary}
\alias{method_view_summary,full_benchmark-method}
\alias{method_view_summary<-}
\alias{method_view_summary<-,full_benchmark,matrix-method}
\title{Retrieve the method_view_summary from a full_benchmark object}
\usage{
method_view_summary(x)

\S4method{method_view_summary}{full_benchmark}(x)

method_view_summary(x) <- value

\S4method{method_view_summary}{full_benchmark,matrix}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{full_benchmark}} object.}

\item{value}{A matrix containing the method_view_summary.}
}
\value{
A matrix containing the method view summary
}
\description{
The \code{method_view_summary} is a matrix containing
summary information from all benchmark runs, organized by run.
}
