% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\docType{class}
\name{musica-class}
\alias{musica-class}
\title{The primary object that contains variants, count_tables,
and samples annotations}
\description{
The primary object that contains variants, count_tables,
and samples annotations
}
\section{Slots}{

\describe{
\item{\code{variants}}{\code{data.table} of variants}

\item{\code{count_tables}}{Summary table with per-sample unnormalized motif counts}

\item{\code{sample_annotations}}{Sample-level annotations (e.g. age, sex, primary)}
}}

