% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/test_data.R
\docType{data}
\name{musica}
\alias{musica}
\alias{musica,musica_result-method}
\title{Retrieve musica from a musica_result object}
\format{
An object of class \code{musica}
See [create_musica()].
}
\usage{
musica(result)

\S4method{musica}{musica_result}(result)

data(musica)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}
}
\value{
A \code{\linkS4class{musica}} musica object
}
\description{
The \code{\linkS4class{musica}} musica contains variants, 
count tables, and sample annotations

A musica created for testing that includes SBS variants
}
\examples{
data(res)
musica(res)
}
\keyword{datasets}
