% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{musica_annot}
\alias{musica_annot}
\title{musica_annot}
\format{
An object of class \code{musica}
See [create_musica_from_variants()] and [create_musica_from_counts()].
}
\usage{
data(musica_annot)
}
\description{
A musica created for testing that includes SBS variants and sample 
annotations
}
\keyword{datasets}
