% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\docType{class}
\name{musica_result-class}
\alias{musica_result-class}
\title{Object containing deconvolved/predicted signatures, sample weights, and
the musica object the result was generated from}
\description{
Object containing deconvolved/predicted signatures, sample weights, and
the musica object the result was generated from
}
\section{Slots}{

\describe{
\item{\code{signatures}}{A matrix of signatures by mutational motifs}

\item{\code{exposures}}{A matrix of samples by signature weights}

\item{\code{table_name}}{A character vector of table names used to make the result}

\item{\code{algorithm}}{Describes how the signatures/weights were generated}

\item{\code{musica}}{The musica object the results were generated from}

\item{\code{umap}}{List of umap data.frames for plotting and analysis}
}}

