% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\docType{class}
\name{musica_result_grid-class}
\alias{musica_result_grid-class}
\title{Object containing the result objects generated from the combination of
annotations and a range of k values}
\description{
Object containing the result objects generated from the combination of
annotations and a range of k values
}
\section{Slots}{

\describe{
\item{\code{grid_params}}{The parameters the result grid was created using}

\item{\code{result_list}}{A list of result objects with different parameters}

\item{\code{grid_table}}{A summary table of the result objects in result_list}
}}

