% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{musica_sbs96}
\alias{musica_sbs96}
\title{musica_sbs96}
\format{
An object of class \code{musica}
See [build_standard_table()].
}
\usage{
data(musica_sbs96)
}
\description{
A musica created for testing that includes SBS variants and a build counts
table for them
}
\keyword{datasets}
