% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{musica_sbs96_tiny}
\alias{musica_sbs96_tiny}
\title{musica_sbs96_tiny}
\format{
An object of class \code{musica}
See [build_standard_table()].
}
\usage{
data(musica_sbs96_tiny)
}
\description{
A very small musica created for testing that includes SBS variants and a
build counts table for them
}
\keyword{datasets}
