% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicatk.R
\name{musicatk}
\alias{musicatk}
\title{Starts the musicatk interactive Shiny app}
\usage{
musicatk(include_version = TRUE, theme = "yeti")
}
\arguments{
\item{include_version}{Include the version number in the header.
Default \code{TRUE}.}

\item{theme}{The theme to use for the GUI. Default \code{"yeti"}.}
}
\value{
The shiny app will open. No data will be returned.
}
\description{
The musicatk Shiny app allows users
to perform mutational signature analysis using an interative graphical user
interface (GUI)
}
\examples{
\dontrun{
# Start the app
musicatk()
}
}
