% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_result_collection.R
\name{parameter}
\alias{parameter}
\alias{parameter,musica-method}
\alias{parameter,result_collection-method}
\alias{parameter<-}
\alias{parameter<-,result_collection,list-method}
\alias{parameter<-,musica,list-method}
\title{Retrieve parameter from a result_collection object}
\usage{
parameter(x, ...)

\S4method{parameter}{musica}(x, result)

\S4method{parameter}{result_collection}(x)

parameter(x, ...) <- value

\S4method{parameter}{result_collection,list}(x) <- value

\S4method{parameter}{musica,list}(x, result) <- value
}
\arguments{
\item{x}{A Object generated by musicatk}

\item{value}{a list of input parameters}
}
\value{
a list of parameters
}
\description{
The \code{parameter} input paramters used in the model
}
