% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_results.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Plot the comparison of two result files}
\usage{
plot_comparison(
  comparison,
  result,
  other_result,
  result_name = deparse(substitute(result)),
  other_result_name = deparse(substitute(other_result)),
  decimals = 2,
  same_scale = TRUE
)
}
\arguments{
\item{comparison}{A matrix detailing the comparison between the two result files.
For example, the output of the \code{compare_results} function.}

\item{result}{A \code{\linkS4class{musica_result}} object.}

\item{other_result}{A second \code{\linkS4class{musica_result}} object.}

\item{result_name}{title for plot of first result signatures}

\item{other_result_name}{title for plot of second result signatures}

\item{decimals}{Specifies rounding for similarity metric displayed. Default
\code{2}.}

\item{same_scale}{If \code{TRUE}, the scale of the probability for each
comparison will be the same. If \code{FALSE}, then the scale of the y-axis
will be adjusted for each comparison. Default \code{TRUE}.}
}
\value{
Returns the comparison plot
}
\description{
Plot the comparison of two result files
}
\examples{
data(res)
comparison <- compare_results(res, res, threshold = 0.8)
plot_comparison(comparison, res, res)
}
