% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{plot_differential_analysis}
\alias{plot_differential_analysis}
\title{Compare exposures of annotated samples}
\usage{
plot_differential_analysis(analysis, analysis_type, samp_num)
}
\arguments{
\item{analysis}{Analysis created by \code{exposure_differential_analysis}}

\item{analysis_type}{Currently only \code{"glm"} supported}

\item{samp_num}{Number of samples that went into the analysis}
}
\value{
Generates a ggplot object
}
\description{
\code{plot_differential_analysis} is used to plot
differential analysis created by \code{exposure_differential_analysis}.
}
\examples{
data("res_annot")
analysis <- exposure_differential_analysis(res_annot, "Tumor_Subtypes", 
method="wilcox")
plot_differential_analysis(analysis, "glm", 2)
}
