% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposures.R
\name{plot_exposure_bar}
\alias{plot_exposure_bar}
\title{Displays sample exposures with bar plots}
\usage{
plot_exposure_bar(
  result,
  proportional = FALSE,
  annotation = NULL,
  num_samples = NULL,
  sort_samples = "total",
  threshold = 0.01,
  same_scale = FALSE,
  label_samples = FALSE,
  legend = TRUE,
  plotly = FALSE
)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{proportional}{If \code{TRUE}, then the exposures will be normalized
to between 0 and 1 by dividing by the total number of counts for each sample.
Default \code{FALSE}.}

\item{annotation}{Sample annotation used to group the subplots. If
\code{NULL}, no grouping by annotation will be performed.
Default \code{NULL}.}

\item{num_samples}{Number of sorted samples to plot}

\item{sort_samples}{This is used to change how samples are sorted in
the barplot from left to right. If set to \code{"total"}, then samples
will be sorted from those with the highest number of mutation counts to the
lowest (regardless of how the parameter \code{"proportional"} is set.
If set to \code{"name"}, then samples are sorted by their name with the
\code{\link[gtools]{mixedsort}} function. If set to 
one or more signature names (e.g. \code{"Signature1"}), then samples will
be sorted from those with the highest level of that signature to the lowest.
If multiple signatures are supplied then,   
Default \code{"total"}.}

\item{threshold}{Exposures less than this threshold will be set to 0
when sorting samples. Only used when \code{sort_samples} is set to the name
of one or more signatures. This is most useful when more than one signature
is supplied as samples that are set to zero for the first exposure will
then be sorted by the levels of the second exposure. Default \code{0.01}.}

\item{same_scale}{If \code{TRUE}, then all subplots will have the 
same scale. Only used when \code{annotation} is supplied. 
Default \code{FALSE}.}

\item{label_samples}{If \code{TRUE}, sample labels will be displayed at
the bottom of the plot. Default \code{FALSE}.}

\item{legend}{If \code{TRUE}, the legend will be displayed.
Default \code{TRUE}.}

\item{plotly}{If \code{TRUE}, the the plot will be made interactive
using \code{\link[plotly]{plotly}}. Default \code{FALSE}.}
}
\value{
Generates a ggplot or plotly object
}
\description{
The distributions of mutational signatures will be displayed 
with bar plots. Samples can be grouped using the \code{annotation} parameter.
Various methods of sorting samples from left to right can be chosen using
the \code{sort_samples} parameter.
}
\examples{
result <- readRDS(system.file("testdata", "res_annot.rds",
package = "musicatk"))
plot_exposure_bar(result, annotation = "Tumor_Subtypes")
}
