% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposures.R
\name{plot_exposure_box}
\alias{plot_exposure_box}
\title{Displays sample exposures with box/violin plots}
\usage{
plot_exposure_box(
  result,
  proportional = FALSE,
  plot_type = c("violin", "box"),
  annotation = NULL,
  group_by = c("none", "annotation", "signature"),
  add_points = TRUE,
  legend = TRUE,
  plotly = FALSE
)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction algorithm.}

\item{proportional}{If \code{TRUE}, then the exposures will be normalized
to between 0 and 1 by dividing by the total number of counts for each sample.
Default \code{FALSE}.}

\item{plot_type}{Determines the type of plot to display. One of 
"violin" or "boxplot". Default \code{violin}.}

\item{annotation}{Sample annotation used to group the subplots. If
\code{NULL}, no grouping by annotation will be performed.
Default \code{NULL}. Only used if the \code{group_by} parameter is set.}

\item{group_by}{Determines how to group the subplots (i.e. facets).
One of \code{"none"}, \code{"signature"} or \code{"annotation"}. If set to
\code{"annotation"}, then a sample annotation must be supplied via the 
\code{annotation} parameter. Default \code{"none"}.}

\item{add_points}{If \code{TRUE}, then points for individual sample 
exposures will be plotted above the violin/box plots. Default \code{TRUE}.}

\item{legend}{If \code{TRUE}, the legend will be displayed.
Default \code{TRUE}.}

\item{plotly}{If \code{TRUE}, the the plot will be made interactive
using \code{\link[plotly]{plotly}}. Default \code{FALSE}.}
}
\value{
Generates a ggplot or plotly object
}
\description{
The distributions of mutational signatures will be displayed 
with box or violin plots (based on the \code{plot_type} parameter).
Samples can be grouped by a sample annotation
saved in the \code{\linkS4class{musica}} object that was used to generate the
\code{\linkS4class{musica_result}} object. Exposures can also be
grouped by signature by changing the \code{group_by} parameter.
}
\examples{
result <- readRDS(system.file("testdata", "res_annot.rds",
package = "musicatk"))
plot_exposure_box(result, group_by = "annotation", annotation = "Tumor_Subtypes")
}
