% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_exposures_by_annotation}
\alias{plot_exposures_by_annotation}
\title{Plots signature weights for each sample gridded by single annotation}
\usage{
plot_exposures_by_annotation(
  result,
  annotation,
  proportional = TRUE,
  label_samples = FALSE,
  sort_samples = "numerical",
  by_group = TRUE,
  num_samples = FALSE,
  no_legend = FALSE,
  thresh_zero = FALSE,
  plotly = FALSE
)
}
\arguments{
\item{result}{S4 Result Object}

\item{annotation}{Annotation to use for facet_wrap}

\item{proportional}{Whether weights are normalized to sum to 1 or not}

\item{label_samples}{Whether to display sample names or not}

\item{sort_samples}{Defaults to numerical, may be total 'counts', or a
specific signatures name (e.g. 'Signature1)}

\item{by_group}{Plot subplots by annotation or by signature}

\item{num_samples}{Number of sorted samples to plot}

\item{no_legend}{Remove legend from the plot}

\item{thresh_zero}{Max level to zero out for better plotting when sorting
by multiple signatures}

\item{plotly}{add plotly layer for plot interaction}
}
\value{
Generates plot {no return}
}
\description{
Plots signature weights for each sample gridded by single annotation
}
\examples{
result <- readRDS(system.file("testdata", "res_annot.rds",
package = "musicatk"))
plot_exposures_by_annotation(result, "Tumor_Subtypes")
}
