% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_full}
\alias{plot_full}
\title{Complete Plotting of input data}
\usage{
plot_full(sample)
}
\arguments{
\item{sample}{Single sample DataFrame}
}
\value{
Generates sample plot {no return}
}
\description{
Complete Plotting of input data
}
