% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmaps using the exposures matrix}
\usage{
plot_heatmap(result, proportional = FALSE, col_annot = FALSE, col_samps, ...)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{proportional}{If \code{TRUE}, then the exposures will be normalized
to between 0 and 1 by dividing by the total number of counts for each sample.
Default \code{FALSE}.}

\item{col_annot}{Boolean check. If \code{True}, column annotations extracted from the musicatk object 
and passed as a HeatmapAnnotation to the Heatmap function.
Default \code{FALSE}.}

\item{col_samps}{Argument containing user given column annotations. It is passed on to the Heatmap function 
as an S4 object of class Heatmap annotation.}

\item{annotations}{Column annotations extracted from the musicatk object}
}
\value{
Generates a heatmap for using the exposure matrix.
}
\description{
The exposures for different signatures can be visualized using a heatmap with this function.
Heatmaps make it easier to visualize the data by representing the magbitude of exposure values
as color in 2-dimensions. The variation in color intensity can help see if the exposures are clustered
or how they vary over space. Exposures can be normalized by providing the \code{proportional} argument.
Column annotations can also be seen by passing the \code{col_annot} argument.
}
\examples{
result <- readRDS(system.file("testdata", "res_annot.rds",
package = "musicatk"))
exp <- exposures(result)
ComplexHeatmap::Heatmap(exp)
}
