% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_val_assit.R
\name{plot_k_comparison}
\alias{plot_k_comparison}
\title{Plot k comparison}
\usage{
plot_k_comparison(k_comparison)
}
\arguments{
\item{k_comparison}{data.frame with k value comparisons returned from the
\code{\link{compare_k_vals}} function.}
}
\value{
a ggplot figure
}
\description{
Plot the results of comparing k values
}
\examples{
data(musica)
k_comparison <- compare_k_vals(musica, "SBS96",
  reps = 3, min_k = 1,
  max_k = 5
)
plot_k_comparison(k_comparison)
}
