% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{plot_Kmeans}
\alias{plot_Kmeans}
\title{Visualize k-means clustering results}
\usage{
plot_Kmeans(result, clusters, group = c("signature", "annotation", "none"))
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool. A two-dimensional UMAP has to
be stored in this object.}

\item{clusters}{The result generated from Kmeans function.}

\item{group}{A single character string indicating the grouping factor. 
Possible options are: "signature" (columns are signatures in a grid), 
"annotation" (columns are sample annotation), and "none" (a single UMAP plot). 
Default is "signature".}
}
\value{
Generate a ggplot object.
}
\description{
The k-means clustering results can be visualized on a UMAP panel. 
Three different types of plots can be generated using this function: cluster-by-signature 
plot, cluster-by-annotation plot, and a single UMAP plot.
}
\examples{
data(res_annot)
#Get k-means result
k_out <- withr::with_seed(123, Kmeans(result = res_annot, centers = 2))
#UMAP
create_umap(result = res_annot)
#generate cluster X signature plot
plot_Kmeans(result = res_annot, clusters = k_out, group = "signature")
#generate cluster X annotation plot
plot_Kmeans(result = res_annot, clusters = k_out, group = "annotation")
#generate a single UMAP plot
plot_Kmeans(result = res_annot, clusters = k_out, group = "none")
}
\seealso{
\link{create_umap}
}
