% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sample_counts}
\alias{plot_sample_counts}
\title{Return sample from musica object}
\usage{
plot_sample_counts(musica, sample_names, table_name = NULL)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{sample_names}{Names or indices of samples to plot.}

\item{table_name}{Name of table used for plotting counts. If \code{NULL},
then the first table in the \code{\linkS4class{musica}} object will be used.
Default \code{NULL}.}
}
\value{
Generates sample plot {no return}
}
\description{
Return sample from musica object
}
\examples{
musica <- readRDS(system.file("testdata", "musica_sbs96.rds", package = "musicatk"))
plot_sample_counts(musica, sample_names = get_sample_names(musica)[1])
}
