% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sample_reconstruction_error}
\alias{plot_sample_reconstruction_error}
\title{Plotting Signature Motif counts/spectra}
\usage{
plot_sample_reconstruction_error(
  result,
  table_name,
  sample_number,
  plotly = FALSE
)
}
\arguments{
\item{result}{S4 Result Object}

\item{table_name}{Table to use for reconstruction error plotting}

\item{sample_number}{Sample number within result to plot error}

\item{plotly}{add plotly layer for plot interaction}
}
\value{
Generates plot {no return}
}
\description{
Plotting Signature Motif counts/spectra
}
\examples{
result <- readRDS(system.file("testdata", "res.rds", package = "musicatk"))
plot_sample_reconstruction_error(result, "SBS96", 1)
}
