% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_signatures}
\alias{plot_signatures}
\title{Plotting Signature Motif counts/spectra}
\usage{
plot_signatures(
  result,
  no_legend = FALSE,
  plotly = FALSE,
  color_variable = NULL,
  color_mapping = NULL,
  text_size = 10,
  facet_size = 10,
  show_x_labels = TRUE
)
}
\arguments{
\item{result}{S4 Result Object}

\item{no_legend}{Remove legend from plot}

\item{plotly}{add plotly layer for plot interaction}

\item{color_variable}{Annotation column to use for coloring plotted motifs,
provided by counts table from input result's musica object}

\item{color_mapping}{Mapping from color_variable to color names, provided by
counts table from input result's musica object}

\item{text_size}{Size of axis text}

\item{facet_size}{Size of facet text}

\item{show_x_labels}{Toggle plotting of x-axis labels}
}
\value{
Generates plot {no return}
}
\description{
Plotting Signature Motif counts/spectra
}
\examples{
result <- readRDS(system.file("testdata", "res.rds", package = "musicatk"))
plot_signatures(result)
}
