% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot a UMAP data.frame}
\usage{
plot_umap(
  result,
  point_size = 0.7,
  legend = TRUE,
  label_clusters = TRUE,
  label_size = 3,
  legend_size = 3,
  text_box = TRUE,
  plotly = FALSE
)
}
\arguments{
\item{result}{Result object containing UMAP data.frame}

\item{point_size}{Scatter plot point size}

\item{legend}{Remove legend}

\item{label_clusters}{Add annotation labels to clusters (may not work well
for split or small clusters)}

\item{label_size}{Size of cluster labels}

\item{legend_size}{Set legend size}

\item{text_box}{Place a box around cluster labels for improved readability}

\item{plotly}{Create plotly version of plot}
}
\value{
Returns a ggplot2 plot of the created umap, if plotly = TRUE the
ggplotly object is returned
}
\description{
Plot a UMAP data.frame
}
\examples{
data(res_annot)
create_umap(res_annot, "Tumor_Subtypes", n_neighbors = 5)
plot_umap(res_annot)
}
