% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_umap_sigs}
\alias{plot_umap_sigs}
\title{Plot a UMAP data.frame}
\usage{
plot_umap_sigs(result)
}
\arguments{
\item{result}{Result object containing UMAP data.frame}
}
\value{
Returns ggplot2 plot of the created umap
}
\description{
Plot a UMAP data.frame
}
\examples{
result <- readRDS(system.file("testdata", "res_annot.rds",
package = "musicatk"))
create_umap(result, "Tumor_Subtypes", n_neighbors = 5)
plot_umap_sigs(result)
}
