% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{predict_and_benchmark}
\alias{predict_and_benchmark}
\title{Predict and benchmark}
\usage{
predict_and_benchmark(
  musica,
  modality,
  k_min,
  k_max,
  full_benchmark,
  algorithm = "lda",
  result_name = "result",
  threshold = 0.8,
  adjustment_threshold = 0.9,
  re_method = "prop",
  plot = FALSE,
  seed = 1,
  nstart = 10,
  par_cores = 1
)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{modality}{Modalitye to use for signature discovery. Needs
to be the same name supplied to the table building functions such as
\link{build_standard_table}.}

\item{k_min}{Minimum number of singatures to predict}

\item{k_max}{Maximum number of signatures to predict}

\item{full_benchmark}{An object of class \code{\linkS4class{full_benchmark}}
created with the \link{create_benchmark} function or returned from a previous
\code{benchmark} run.}

\item{algorithm}{Method to use for mutational signature discovery. One of 
\code{"lda"} or \code{"nmf"}. Default \code{"lda"}.}

\item{result_name}{Name of the result list entry to save prediction. 
Default \code{"result"}.}

\item{threshold}{Cosine similarity cutoff for comparing predicted and true
signatures. Default \code{0.8}.}

\item{adjustment_threshold}{Cosine similarity value of high confidence.
Comparisons that meet this cutoff are assumed to be likely,
while those that fall below the cutoff will be disregarded if the predicted
signature is already captured above the threshold. Default \code{0.9}.}

\item{re_method}{Method for reconstruction error calculation, either
proportion or raw. Default \code{"prop"}.}

\item{plot}{If \code{FALSE}, plots will be suppressed. Default \code{TRUE}.}

\item{seed}{Seed to be used for the random number generators in the
signature discovery algorithms. Default \code{1}.}

\item{nstart}{Number of independent random starts used in the mutational
signature algorithms. Default \code{10}.}

\item{par_cores}{Number of parallel cores to use. Only used if
\code{method = "nmf"}. Default \code{1}.}
}
\value{
If \code{make_copy == TRUE}, a new \code{full_benchmark} object is
returned. If \code{make_copy == FALSE}, nothing is returned.
}
\description{
This function will discover signatures from a musica object using a given
algorithm and a range of k values. A new discovery is done for each k value.
As each discovery is completed, the prediction is benchmarked.
}
\examples{
data(synthetic_breast_counts)
data(synthetic_breast_true_exposures)
true_sigs <- c("SBS1", "SBS2", "SBS3", "SBS8", "SBS13", "SBS17", "SBS18", "SBS26")
true_signatures <- signatures(cosmic_v2_sigs)[,true_sigs]
full_benchmark <- create_benchmark(true_signatures, 
  synthetic_breast_true_exposures, synthetic_breast_counts, "SBS96")
prediction_musica <- create_musica_from_counts(synthetic_breast_counts, "SBS96")
predict_and_benchmark(prediction_musica, "SBS96", 7, 9, full_benchmark, "nmf")

}
