% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_prediction.R
\name{predict_exposure}
\alias{predict_exposure}
\title{LDA prediction of samples based on existing signatures}
\usage{
predict_exposure(
  musica,
  g,
  table_name,
  signature_res,
  algorithm,
  signatures_to_use = seq_len(ncol(signature_res@signatures)),
  seed = 1,
  verbose = FALSE
)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica_data}} object.}

\item{g}{A \linkS4class{BSgenome} object indicating which genome
reference the variants and their coordinates were derived from.}

\item{table_name}{Name of table used for posterior prediction.
Must match the table type used to generate the prediction signatures}

\item{signature_res}{Signatures to use for prediction}

\item{algorithm}{Algorithm to use for prediction. Choose from
"lda_posterior", decompTumor2Sig, and deconstructSigs}

\item{signatures_to_use}{Which signatures in set to use (default all)}

\item{seed}{Seed to use for reproducible results, set to null to disable}

\item{verbose}{Whether to show intermediate results}
}
\value{
A result object containing signatures and sample weights
}
\description{
LDA prediction of samples based on existing signatures
}
\examples{
musica <- readRDS(system.file("testdata", "musica.rds", package = "musicatk"))
g <- select_genome("19")
build_standard_table(musica, g, "SBS96", overwrite = TRUE)
predict_exposure(musica = musica, table_name = "SBS96",
signature_res = cosmic_v2_sigs, algorithm = "lda")
}
