% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{rc}
\alias{rc}
\title{Reverse complement of a string using biostrings}
\usage{
rc(dna)
}
\arguments{
\item{dna}{Input DNA string}
}
\value{
Returns the reverse compliment of the input DNA string
}
\description{
Reverse complement of a string using biostrings
}
\examples{
rc("ATGC")
}
