% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic_data.R
\docType{data}
\name{rep_range}
\alias{rep_range}
\title{Replication Timing Data as GRanges Object}
\format{
An object of class \code{"GRanges"}; see
[annotate_replication_strand()].
}
\source{
GEO, <https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE134225>
}
\usage{
data(rep_range)
}
\description{
Supplementary data converted from bigWig to bedgraph to GRanges, with low
RFD indicating the leading strand and high RFD indicating lagging strand and
removing uninformative zero RFD intervals. Timing data is 10kb bins from
a colon cancer sample.
}
\references{
Sriramachandran, A. M. et al. (2020)
Genome-wide Nucleotide-Resolution Mapping of DNA Replication Patterns,
Single-Strand Breaks, and Lesions by GLOE-Seq. ([Molecular Cell]
(doi:10.1016/j.molcel.2020.03.027))
}
\keyword{datasets}
