% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{res}
\alias{res}
\title{res}
\format{
An object of class \code{musica}
See [discover_signatures()].
}
\usage{
data(res)
}
\description{
A musica created for testing that includes SBS variants with discovered
exposures and signatures
}
\keyword{datasets}
