% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{res_annot}
\alias{res_annot}
\title{res_annot}
\format{
An object of class \code{musica}
See [discover_signatures()].
}
\usage{
data(res_annot)
}
\description{
A musica created for testing that includes SBS variants with annotations and
discovered exposures and signatures
}
\keyword{datasets}
