% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_result_collection.R
\docType{class}
\name{result_collection-class}
\alias{result_collection-class}
\title{The Result Collection object that contains modality, input parameters, prior
hyperparameters}
\description{
The Result Collection object that contains modality, input parameters, prior
hyperparameters
}
\section{Slots}{

\describe{
\item{\code{modality}}{a list contains model results for different modality}

\item{\code{parameter}}{a list contains input parameters}

\item{\code{hyperparameter}}{a list contains prior and tuning parameters}
}}

