% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_musica.R
\name{result_list}
\alias{result_list}
\alias{result_list,musica-method}
\alias{result_list<-}
\alias{result_list<-,musica,list-method}
\title{Retrieve result_list from a musica object}
\usage{
result_list(object)

\S4method{result_list}{musica}(object)

result_list(musica) <- value

\S4method{result_list}{musica,list}(musica) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica} function.}

\item{musica}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica} function}

\item{value}{A list of results}
}
\value{
A list of results
}
\description{
The \code{result_list} contains results from various runs
}
\examples{
data(res)
result_list(res)
data(musica)
result_list(musica)
}
