% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_result_model.R
\docType{class}
\name{result_model-class}
\alias{result_model-class}
\title{Object that contains results for a single model}
\description{
Object that contains results for a single model
}
\section{Slots}{

\describe{
\item{\code{signatures}}{A matrix of signatures by mutational motifs}

\item{\code{exposures}}{A matrix of samples by signature weights}

\item{\code{num_signatures}}{Number of signatures in the model}

\item{\code{other_parameters}}{Parameters relevant to the model}

\item{\code{credible_intervals}}{Credible intervals for parameters}

\item{\code{metrics}}{Performance metrics for the model}

\item{\code{umap}}{List of umap data.frames for plotting and analysis}

\item{\code{model_id}}{Model identifier}

\item{\code{modality}}{Modality of result (SBS96, DBS78, IND83)}
}}

