% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sample_annotations}
\alias{sample_annotations}
\alias{sample_annotations,musica-method}
\alias{sample_annotations,musica_result-method}
\alias{sample_annotations<-}
\alias{sample_annotations<-,musica,character,vector-method}
\alias{sample_annotations<-,musica_result,character,vector-method}
\title{Get or set sample annotations from a musica or musica_result object}
\usage{
sample_annotations(object)

\S4method{sample_annotations}{musica}(object)

\S4method{sample_annotations}{musica_result}(object)

sample_annotations(object, name) <- value

\S4method{sample_annotations}{musica,character,vector}(object, name) <- value

\S4method{sample_annotations}{musica_result,character,vector}(object, name) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica} function or a \code{\linkS4class{musica_result}}
object generated by a mutational discovery or prediction tool.}

\item{name}{The name of the new annotation to add.}

\item{value}{A vector containing the new sample annotations. Needs to be
the same length as the number of samples in the object.}
}
\value{
A new object with the sample annotations added to the table in the 
\code{sample_annotations} slot.
}
\description{
Sample annotations can be used to store information about
each sample such as tumor type or treatment status. These are used in 
downstream plotting functions such as \code{\link{plot_exposures}} or 
\code{\link{plot_umap}} to group or color samples by a particular annotation.
}
\examples{
result <- readRDS(system.file("testdata", "res_annot.rds", package = "musicatk"))
sample_annotations(result)

# Add new annotation
sample_annotations(result, "New_Annotation") <- rep(c("A", "B"), c(3, 4))
sample_annotations(result)
}
\seealso{
See \code{\link{get_sample_names}} to get the list of samples in 
the \code{\linkS4class{musica}} or \code{\linkS4class{musica_result}} object.
}
