% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{set_sample_annotations}
\alias{set_sample_annotations}
\title{Set sample level annotations for musica object}
\usage{
set_sample_annotations(musica, annotations)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{annotations}{Sample DataFrame}
}
\value{
Sets sample_annotations slot {no return}
}
\description{
Set sample level annotations for musica object
}
\examples{
musica <- readRDS(system.file("testdata", "musica_sbs96.rds", package = "musicatk"))
sample_annotations <- read.table(system.file("testdata",
"sample_annotations.txt", package = "musicatk"), sep = "\t", header=TRUE)
set_sample_annotations(musica, sample_annotations)
}
