% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_full_benchmark.R
\name{sig_view_summary}
\alias{sig_view_summary}
\alias{sig_view_summary,full_benchmark-method}
\alias{sig_view_summary<-}
\alias{sig_view_summary<-,full_benchmark,matrix-method}
\title{Retrieve the sig_view_summary from a full_benchmark object}
\usage{
sig_view_summary(x)

\S4method{sig_view_summary}{full_benchmark}(x)

sig_view_summary(x) <- value

\S4method{sig_view_summary}{full_benchmark,matrix}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{full_benchmark}} object.}

\item{value}{A matrix containing the sig_view_summary}
}
\value{
A matrix containing the signature view summary
}
\description{
The \code{sig_view_summary} is a matrix containing
summary information from all benchmark runs, organized by signature.
}
