% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\docType{class}
\name{single_benchmark-class}
\alias{single_benchmark-class}
\title{Object that contains information for an individual benchmark run}
\description{
Object that contains information for an individual benchmark run
}
\section{Slots}{

\describe{
\item{\code{initial_pred}}{The initial prediction}

\item{\code{intermediate_pred}}{The intermediate prediction, after duplicates are
corrected, but before composites are corrected}

\item{\code{final_pred}}{The final prediction}

\item{\code{initial_comparison}}{The initial comparison between predicted and true
signatures}

\item{\code{intermediate_comparison}}{The intermediate comparison between predicted
and true signatures}

\item{\code{final_comparison}}{The final comparison between predicted and true
signatures}

\item{\code{single_summary}}{A matrix summarizing the benchmark run}

\item{\code{method_id}}{An identifier for the benchmark run}

\item{\code{threshold}}{Cosine similarity cutoff for comparing predicted and true
signatures}

\item{\code{adjustment_threshold}}{Cosine similarity value of high confidence.
Comparisons that meet this cutoff are assumed to be likely,
while those that fall below the cutoff will be disregarded if the predicted
signature is already captured above the threshold.}

\item{\code{description}}{Further details about the prediction being benchmarked.}
}}

