% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{single_summary}
\alias{single_summary}
\alias{single_summary,single_benchmark-method}
\alias{single_summary<-}
\alias{single_summary<-,single_benchmark,matrix-method}
\title{Retrieve the single summary matrix from a single_benchmark object}
\usage{
single_summary(x)

\S4method{single_summary}{single_benchmark}(x)

single_summary(x) <- value

\S4method{single_summary}{single_benchmark,matrix}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark}} object.}

\item{value}{A matrix containing the summary infomation for the run.}
}
\value{
A matrix containing the summary.
}
\description{
The \code{single_summary} contains the summary information
from the benchmark run.
}
