% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{subset_musica_by_annotation}
\alias{subset_musica_by_annotation}
\title{Creates a new musica object subsetted to only one value of a sample annotation}
\usage{
subset_musica_by_annotation(musica, annot_col, annot_names)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica_data}} object.}

\item{annot_col}{Annotation class to use for subsetting}

\item{annot_names}{Annotational value to subset to}
}
\value{
Returns a new musica object with sample annotations, count tables,
and variants subsetted to only contains samples of the specified annotation
type
}
\description{
Creates a new musica object subsetted to only one value of a sample annotation
}
\examples{
musica <- readRDS(system.file("testdata", "musica_sbs96.rds", package = "musicatk"))
sample_annotations <- read.table(system.file("testdata",
"sample_annotations.txt", package = "musicatk"), sep = "\t", header=TRUE)
init_sample_annotations(musica)
add_sample_annotations(musica, sample_annotations, sample_column =
"Sample_Names", columns_to_add = "Tumor_Subtypes")
subset_musica_by_annotation(musica, "Tumor_Subtypes", "Lung")
}
