% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{subset_variant_by_type}
\alias{subset_variant_by_type}
\title{Subsets a variant table based on Variant Type}
\usage{
subset_variant_by_type(tab, type)
}
\arguments{
\item{tab}{Input variant table}

\item{type}{Variant type to return e.g. "SBS", "INS", "DEL", "DBS"}
}
\value{
Returns the input variant table subsetted to only contain variants
of the specified variant type
}
\description{
Subsets a variant table based on Variant Type
}
\examples{
data(musica)
annotate_variant_type(musica)
subset_variant_by_type(get_variants(musica), "SBS")
}
