% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{synthetic_breast_counts}
\alias{synthetic_breast_counts}
\title{synthetic_breast_counts}
\format{
An object of class \code{data.frame}
}
\usage{
data(synthetic_breast_counts)
}
\description{
A data.frame containing the SBS96 mutation counts for a synthetic breast 
cancer dataset with 214 samples.
}
\keyword{datasets}
