% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{synthetic_breast_true_exposures}
\alias{synthetic_breast_true_exposures}
\title{synthetic_breast_true_exposures}
\format{
An object of class \code{data.frame}
}
\usage{
data(synthetic_breast_true_exposures)
}
\description{
A data.frame containing true signature exposure levels for a synthetic 
breast cancer dataset with 214 samples and 8 signatures.
}
\keyword{datasets}
