% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{table_96}
\alias{table_96}
\title{Generates a 96 motif table based on input counts for plotting}
\usage{
table_96(sample_df)
}
\arguments{
\item{sample_df}{Input counts table}
}
\value{
Returns a 96 motif summary table
}
\description{
Generates a 96 motif table based on input counts for plotting
}
