% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{table_name}
\alias{table_name}
\alias{table_name,musica_result-method}
\title{Retrieve table name used for plotting from a musica_result object}
\usage{
table_name(result)

\S4method{table_name}{musica_result}(result)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}
}
\value{
Table name used for plotting
}
\description{
The table name
}
\examples{
data(res)
table_name(res)
}
