% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{table_selected}
\alias{table_selected}
\alias{table_selected,musica_result-method}
\title{Retrieve table name used for plotting from a musica_result object}
\usage{
table_selected(result)

\S4method{table_selected}{musica_result}(result)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}
}
\value{
Table name used for plotting
}
\description{
The table name
}
\examples{
data(res)
table_selected(res)
}
