% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tables}
\alias{tables}
\alias{tables,musica-method}
\alias{tables,musica_result-method}
\alias{tables<-}
\alias{tables<-,musica,list-method}
\title{Retrieve the list of count_tables from a musica or musica_result 
object}
\usage{
tables(object)

\S4method{tables}{musica}(object)

\S4method{tables}{musica_result}(object)

tables(musica) <- value

\S4method{tables}{musica,list}(musica) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica_from_variants} or \link{create_musica_from_counts} function,
or a \code{\linkS4class{musica_result}}
object generated by a mutational discovery or prediction tool.}

\item{musica}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica_from_variants} or \link{create_musica_from_counts} function, 
or a \code{\linkS4class{musica_result}}
object generated by a mutational discovery or prediction tool.}

\item{value}{A list of \code{\linkS4class{count_table}} objects representing 
counts of motifs in samples}
}
\value{
A list of count_tables
}
\description{
The \code{count_tables} contains standard and/or custom 
count tables created from variants
}
\examples{
data(res)
tables(res)
data(musica)
tables(musica)
}
