% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_single_benchmark.R
\name{threshold}
\alias{threshold}
\alias{threshold,single_benchmark-method}
\alias{threshold<-}
\alias{threshold<-,single_benchmark,numeric-method}
\title{Retrieve the threshold from a single_benchmark object}
\usage{
threshold(x)

\S4method{threshold}{single_benchmark}(x)

threshold(x) <- value

\S4method{threshold}{single_benchmark,numeric}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{single_benchmark}} object.}

\item{value}{The threshold}
}
\value{
The threshold value.
}
\description{
The \code{threshold} is the cosine similarity cutoff for
comparing predicted and true signatures.
}
