% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{adjustment_threshold}
\alias{adjustment_threshold}
\alias{adjustment_threshold,single_benchmark-method}
\alias{adjustment_threshold<-}
\alias{adjustment_threshold<-,single_benchmark,numeric-method}
\title{Retrieve the adjustment_threshold from a single_benchmark object}
\usage{
adjustment_threshold(result)

\S4method{adjustment_threshold}{single_benchmark}(result)

adjustment_threshold(result) <- value

\S4method{adjustment_threshold}{single_benchmark,numeric}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{single_benchmark}} object extracted from
a \code{\linkS4class{full_benchmark}} object}

\item{value}{The adjustment_threshold for the single benchmark run}
}
\value{
The adjustment_threshold for the single benchmark run
}
\description{
The \code{adjustment_threshold} is the cosine similarity value
of high confidence. Comparisons that meet this cutoff are assumed to be likely,
while those that fall below the cutoff will be disregarded if the predicted
signature is already captured above the threshold.
}
