% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{benchmark}
\alias{benchmark}
\title{Run the benchmark framework on a prediction}
\usage{
benchmark(
  full_benchmark,
  prediction,
  method_id = NULL,
  threshold = 0.8,
  adjustment_threshold = 0.9,
  description = NULL,
  plot = TRUE,
  make_copy = FALSE
)
}
\arguments{
\item{full_benchmark}{An object of class \code{\linkS4class{full_benchmark}}
created with the \link{create_benchmark} function or returned from a previous
\code{benchmark} run.}

\item{prediction}{An object of class \code{\linkS4class{musica_result}}
containing the predicted signatures and exposures to benchmark.}

\item{method_id}{An identifier for the prediction being benchmarked. If not
supplied, it will be automatically set to the variable name of the prediction
provided. Default \code{NULL}.}

\item{threshold}{Cosine similarity cutoff for comparing preidcted and true
signatures. Default \code{0.8}.}

\item{adjustment_threshold}{Cosine similarity value of high confidence.
Comparisons that meet this cutoff are assumed to be likely,
while those that fall below the cutoff will be disregarded if the predicted
signature is already captured above the threshold. Default \code{0.9}.}

\item{description}{Further details about the prediction being benchmarked.
Default \code{NULL}.}

\item{plot}{If \code{FALSE}, plots will be suppressed. Default \code{TRUE}.}

\item{make_copy}{If \code{TRUE}, the \code{full_benchmark} object provided
will not be modified and a new object will be returned. If \code{FALSE}, the
object provided will be modified and nothing will be returned. Default
\code{FALSE}.}
}
\value{
If \code{make_copy == TRUE}, a new \code{full_benchmark} object is
returned. If \code{make_copy == FALSE}, nothing is returned.
}
\description{
Perform benchmarking on a signature discovery prediction compared
to a ground truth. Potential errors in the predicted signatures, such as
composite or duplicate signatures are adjusted, and a summary of the accuracy
of the prediction is given.
}
