% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{benchmark_plot_comparison}
\alias{benchmark_plot_comparison}
\title{Plot a signature comparison from a benchmarking analysis}
\usage{
benchmark_plot_comparison(
  full_benchmark,
  method_id,
  prediction,
  decimals = 2,
  same_scale = FALSE
)
}
\arguments{
\item{full_benchmark}{The \code{\linkS4class{full_benchmark}} object for the
benchmarking analysis}

\item{method_id}{The identifier for the \code{\linkS4class{single_benchmark}}
object containing the comparison of interest}

\item{prediction}{\code{Initial} for the comparison before any benchmarking
adjustments have been made, \code{Intermediate} for the comparison after
duplicates have been adjusted but before composites are adjusted, or
\code{Final} for the comparison at the end of the benchmarking adjustments.}

\item{decimals}{Specifies rounding for similarity metric displayed. Default
\code{2}.}

\item{same_scale}{If \code{TRUE}, the scale of the probability for each
comparison will be the same. If \code{FALSE}, then the scale of the y-axis
will be adjusted for each comparison. Default \code{TRUE}.}
}
\value{
Returns the comparison plot
}
\description{
After a prediction has been benchmarked with the \link{benchmark} function,
the comparison between the true and predicted signatures at any step of the
benchmarking process can be plotted.
}
