% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{benchmark_plot_duplicate_exposures}
\alias{benchmark_plot_duplicate_exposures}
\title{Plot effect of duplicate correction}
\usage{
benchmark_plot_duplicate_exposures(full_benchmark, method_id)
}
\arguments{
\item{full_benchmark}{The \code{\linkS4class{full_benchmark}} object for the
benchmarking analysis}

\item{method_id}{The identifier for the \code{\linkS4class{single_benchmark}}
object of interest}
}
\value{
Two ggplot images are displayed
}
\description{
After a prediction has been benchmarked with the \link{benchmark} function,
the true and predicted exposures can be plotted both before and after the
duplicate signature adjustment. The effect of the adjustment can then be
observed.
}
