% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{benchmark_plot_signatures}
\alias{benchmark_plot_signatures}
\title{Plot signatures from a benchmarking analysis}
\usage{
benchmark_plot_signatures(
  full_benchmark,
  method_id,
  prediction,
  plotly = FALSE,
  color_variable = NULL,
  color_mapping = NULL,
  text_size = 10,
  show_x_labels = TRUE,
  show_y_labels = TRUE,
  same_scale = TRUE,
  y_max = NULL,
  annotation = NULL,
  percent = TRUE
)
}
\arguments{
\item{full_benchmark}{The \code{\linkS4class{full_benchmark}} object for the
benchmarking analysis}

\item{method_id}{The identifier for the \code{\linkS4class{single_benchmark}}
object containing the signatures to be plotted}

\item{prediction}{\code{Initial} for the signatures before any benchmarking
adjustments have been made, \code{Intermediate} for the signatures after
duplicates have been adjusted but before composites are adjusted, or
\code{Final} for the signatures at the end of the benchmarking adjustments.}

\item{plotly}{If \code{TRUE}, the the plot will be made interactive
using \code{\link[plotly]{plotly}}. Default \code{FALSE}.}

\item{color_variable}{Name of the column in the variant annotation data.frame
to use for coloring the mutation type bars. The variant annotation data.frame 
can be found within the count table of the \code{\linkS4class{musica}}
object. If \code{NULL}, then the default column specified in the count
table will be used. Default \code{NULL}.}

\item{color_mapping}{A character vector used to map items in the
\code{color_variable} to a color. The items in \code{color_mapping}
correspond to the colors. The names of the items in \code{color_mapping}
should correspond to the uniqeu items in \code{color_variable}. If
\code{NULL}, then the default \code{color_mapping} specified in the count
table will be used. Default \code{NULL}.}

\item{text_size}{Size of axis text. Default \code{10}.}

\item{show_x_labels}{If \code{TRUE}, the labels for the mutation types
on the x-axis will be shown. Default \code{TRUE}.}

\item{show_y_labels}{If \code{TRUE}, the y-axis ticks and labels will be 
shown. Default \code{TRUE}.}

\item{same_scale}{If \code{TRUE}, the scale of the probability for each
signature will be the same. If \code{FALSE}, then the scale of the y-axis
will be adjusted for each signature. Default \code{TRUE}.}

\item{y_max}{Vector of maximum y-axis limits for each signature. One value 
may also be provided to specify a constant y-axis limit for all signatures.
Vector length must be 1 or equivalent to the number of signatures. Default 
\code{NULL}.}

\item{annotation}{Vector of annotations to be displayed in the top right
corner of each signature. Vector length must be equivalent to the number of
signatures. Default \code{NULL}.}

\item{percent}{If \code{TRUE}, the y-axis will be represented in percent 
format instead of mutation counts. Default \code{TRUE}.}
}
\value{
Generates a ggplot or plotly object
}
\description{
After a prediction has been benchmarked with the \link{benchmark} function,
this function can be used to plot signatures from any step in the benchmarking
process. Comparable to the \code{plot_signatures} function but compatible with
benchmarking objects.
}
