% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{build_standard_table}
\alias{build_standard_table}
\title{Builds a standard table from user variants}
\usage{
build_standard_table(
  musica,
  g,
  table_name,
  strand_type = NA,
  overwrite = FALSE
)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{g}{A \linkS4class{BSgenome} object indicating which genome
reference the variants and their coordinates were derived from.}

\item{table_name}{Name of standard table to build SBS96, SBS192, DBS, or
Indel}

\item{strand_type}{Only for SBS192 Transcript_Strand or Replication_Strand}

\item{overwrite}{Overwrite existing count table}
}
\value{
None
}
\description{
Builds a standard table from user variants
}
\examples{
g <- select_genome("19")

data(musica)
build_standard_table(musica, g, "SBS96", overwrite = TRUE)

data(musica)
annotate_transcript_strand(musica, "19")
build_standard_table(musica, g, "SBS192", "Transcript_Strand")

data(musica)
data(rep_range)
annotate_replication_strand(musica, rep_range)
build_standard_table(musica, g, "SBS192", "Replication_Strand")

data(dbs_musica)
build_standard_table(dbs_musica, g, "DBS")

data(indel_musica)
build_standard_table(indel_musica, g, table_name = "INDEL")
}
