% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{built_tables}
\alias{built_tables}
\alias{built_tables,musica-method}
\alias{built_tables,musica_result-method}
\title{Retrieve the names of count_tables from a musica or musica_result
object}
\usage{
built_tables(object)

\S4method{built_tables}{musica}(object)

\S4method{built_tables}{musica_result}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica_from_variants} or \link{create_musica_from_counts} function,
or a \code{\linkS4class{musica_result}}
object generated by a mutational discovery or prediction tool.}
}
\value{
The names of created count_tables
}
\description{
The \code{count_tables} contains standard and/or custom
count tables created from variants
}
\examples{
data(res)
built_tables(res)
}
