% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_prediction.R
\name{combine_predict_grid}
\alias{combine_predict_grid}
\title{Combine prediction grid list into a result object. Exposure values are zero
for samples in an annotation where that signature was not predicted}
\usage{
combine_predict_grid(grid_list, musica, signature_res)
}
\arguments{
\item{grid_list}{A list of result objects from the prediction grid to
combine into a single result}

\item{musica}{A \code{\linkS4class{musica}} object.}

\item{signature_res}{Signatures to automatically subset from for prediction}
}
\value{
A result object combining all samples and signatures from a
prediction grid. Samples have zero exposure value for signatures not found
in that annotation type.
}
\description{
Combine prediction grid list into a result object. Exposure values are zero
for samples in an annotation where that signature was not predicted
}
\examples{
musica <- readRDS(system.file("testdata", "musica_annot.rds", package = "musicatk"))
grid <- auto_predict_grid(musica, "SBS96", cosmic_v2_sigs, "lda",
"Tumor_Subtypes", combine_res = FALSE)
combined <- combine_predict_grid(grid, musica, cosmic_v2_sigs)
plot_exposures_by_annotation(combined, "Tumor_Subtypes")
}
