% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_results.R
\name{compare_cosmic_v2}
\alias{compare_cosmic_v2}
\title{Compare a result object to COSMIC V2 SBS Signatures (combination whole-exome
and whole-genome)}
\usage{
compare_cosmic_v2(
  result,
  threshold = 0.9,
  metric = "cosine",
  result_name = deparse(substitute(result)),
  show_plot = TRUE
)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object.}

\item{threshold}{threshold for similarity}

\item{metric}{One of \code{"cosine"} for cosine similarity or \code{"jsd"} 
for 1 minus the Jensen-Shannon Divergence. Default \code{"cosine"}.}

\item{result_name}{title for plot user result signatures}

\item{show_plot}{Toggle display of plot, or only output table}
}
\value{
Returns the comparisons
}
\description{
Compare a result object to COSMIC V2 SBS Signatures (combination whole-exome
and whole-genome)
}
\examples{
data(res)
compare_cosmic_v2(res, threshold = 0.7)
}
