% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_results.R
\name{compare_results}
\alias{compare_results}
\title{Compare two result files to find similar signatures}
\usage{
compare_results(
  result,
  other_result,
  threshold = 0.9,
  metric = "cosine",
  type = "threshold",
  result_name = deparse(substitute(result)),
  other_result_name = deparse(substitute(other_result))
)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object.}

\item{other_result}{A second \code{\linkS4class{musica_result}} object.}

\item{threshold}{threshold for similarity}

\item{metric}{One of \code{"cosine"} for cosine similarity or \code{"jsd"} 
for 1 minus the Jensen-Shannon Divergence. Default \code{"cosine"}.}

\item{type}{One of \code{"threshold"} for all results above threshold or 
\code{"best"} for the best match results above threshold}

\item{result_name}{title for plot of first result signatures}

\item{other_result_name}{title for plot of second result signatures}
}
\value{
Returns the comparisons
}
\description{
Compare two result files to find similar signatures
}
\examples{
data(res)
compare_results(res, res, threshold = 0.8)
}
