% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_results.R
\name{compare_results}
\alias{compare_results}
\title{Compare two result files to find similar signatures}
\usage{
compare_results(result, other_result, threshold = 0.9, metric = "cosine")
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object.}

\item{other_result}{A second \code{\linkS4class{musica_result}} object.}

\item{threshold}{threshold for similarity}

\item{metric}{One of \code{"cosine"} for cosine similarity or \code{"jsd"} 
for 1 minus the Jensen-Shannon Divergence. Default \code{"cosine"}.}
}
\value{
Returns the comparisons
}
\description{
Compare two result files to find similar signatures
}
\examples{
data(res)
compare_results(res, res, threshold = 0.8)
}
