% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{create_benchmark}
\alias{create_benchmark}
\title{Create a full_benchmark object}
\usage{
create_benchmark(true_signatures, true_exposures, count_table)
}
\arguments{
\item{true_signatures}{A matrix of true signatures by mutational motifs}

\item{true_exposures}{A matrix of samples by true signature weights}

\item{count_table}{Summary table with per-sample unnormalized motif counts}
}
\value{
A \code{\linkS4class{full_benchmark}} object
}
\description{
Initialize a \code{\linkS4class{full_benchmark}} object for benchmarking
}
