% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{create_musica}
\alias{create_musica}
\title{Creates a musica object from a variant table}
\usage{
create_musica(
  x,
  genome,
  check_ref_chromosomes = TRUE,
  check_ref_bases = TRUE,
  chromosome_col = "chr",
  start_col = "start",
  end_col = "end",
  ref_col = "ref",
  alt_col = "alt",
  sample_col = "sample",
  extra_fields = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{Any object that can be coerced to a data.table including a matrix
or data.frame.}

\item{genome}{A \linkS4class{BSgenome} object indicating which genome
reference the variants and their coordinates were derived from.}

\item{check_ref_chromosomes}{Whether to peform a check to ensure that
the chromosomes in the \code{variant} object match the reference
chromosomes in the \code{genome} object. If there are mismatches, this
may cause errors in downstream generation of count tables. If mismatches
occur, an attept to be automatically fix these with the
\code{\link[GenomeInfoDb]{seqlevelsStyle}} function will be made.
Default \code{TRUE}.}

\item{check_ref_bases}{Whether to check if the reference bases in the
\code{variant} object match the reference bases in the \code{genome}
object. Default \code{TRUE}.}

\item{chromosome_col}{The name of the column that contains the chromosome
reference for each variant. Default \code{"Chromosome"}.}

\item{start_col}{The name of the column that contains the start
position for each variant. Default \code{"Start_Position"}.}

\item{end_col}{The name of the column that contains the end
position for each variant. Default \code{"End_Position"}.}

\item{ref_col}{The name of the column that contains the reference
base(s) for each variant. Default \code{"Tumor_Seq_Allele1"}.}

\item{alt_col}{The name of the column that contains the alternative
base(s) for each variant. Default \code{"Tumor_Seq_Allele2"}.}

\item{sample_col}{The name of the column that contains the sample
id for each variant. Default \code{"Tumor_Sample_Barcode"}.}

\item{extra_fields}{Which additional fields to extract and include in
the musica object. Default \code{NULL}.}

\item{verbose}{Whether to print status messages during error checking.
Default \code{TRUE}.}
}
\value{
Returns a musica object
}
\description{
Add description
}
\examples{
maf_file <- system.file("testdata", "public_TCGA.LUSC.maf",
package = "musicatk")
variants <- extract_variants_from_maf_file(maf_file)
g <- select_genome("38")
musica <- create_musica(x = variants, genome = g)
}
