% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{create_umap}
\alias{create_umap}
\title{Create a UMAP from a musica result}
\usage{
create_umap(result, n_neighbors = 30, min_dist = 0.75, spread = 1)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{n_neighbors}{The size of local neighborhood used for views of
manifold approximation. Larger values result in more global the manifold,
while smaller values result in more local data being preserved.
If \code{n_neighbors} is larger than the number of samples,
then \code{n_neighbors} will automatically be set to the number of samples
in the \code{\linkS4class{musica_result}}. Default \code{30}.}

\item{min_dist}{The effective minimum distance between embedded points.
Smaller values will result in a more clustered/clumped embedding where
nearby points on the manifold are drawn closer together, while larger
values will result on a more even dispersal of points. Default \code{0.2}.}

\item{spread}{The effective scale of embedded points. In combination with
‘min_dist’, this determines how clustered/clumped the embedded points are.
Default \code{1}.}
}
\value{
A \code{\linkS4class{musica_result}} object with a new UMAP
stored in the \code{UMAP} slot.
}
\description{
Proportional sample exposures will be used as input into the 
\code{\link[uwot]{umap}} function to generate a two dimensional UMAP.
}
\examples{
#data(res_annot)
#create_umap(result = res_annot)
}
\seealso{
See \link{plot_umap} to display the UMAP and
\code{\link[uwot]{umap}} for more information on the individual parameters
for generating UMAPs.
}
