% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{description}
\alias{description}
\alias{description,single_benchmark-method}
\alias{description<-}
\alias{description<-,single_benchmark,character-method}
\title{Retrieve the description from a single_benchmark object}
\usage{
description(result)

\S4method{description}{single_benchmark}(result)

description(result) <- value

\S4method{description}{single_benchmark,character}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{single_benchmark}} object extracted from
a \code{\linkS4class{full_benchmark}} object}

\item{value}{The description for the single benchmark run}
}
\value{
The description for the single benchmark run
}
\description{
The \code{description} contains details about the prediction
that is benchmarked.
}
