% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{.jsd}
\alias{.jsd}
\title{Calculates 1 - Jensen-Shannon Divergences between all pairs of columns
between two matrices}
\usage{
.jsd(p, q, epsilon = 1e-07)
}
\arguments{
\item{p}{First matrix}

\item{q}{Second matrix}

\item{epsilon}{Number to add to all probabilities. Default \code{0.0000001}.}
}
\value{
Returns matrix of 1 - Jensen-Shannon Divergences
}
\description{
Calculates 1 - Jensen-Shannon Divergences between all pairs of columns
between two matrices
}
\keyword{internal}
