% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{exposures}
\alias{exposures}
\alias{exposures,musica_result-method}
\title{Retrieve exposures from a musica_result object}
\usage{
exposures(result)

\S4method{exposures}{musica_result}(result)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}
}
\value{
A matrix of exposures
}
\description{
The \code{exposure} matrix contains estimated amount of
each signature for each sample. Rows correspond to each signature and
columns correspond to each sample.
}
\examples{
result <- readRDS(system.file("testdata", "res.rds", package = "musicatk"))
exposures(result)
}
