% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{exposures}
\alias{exposures}
\alias{exposures,musica_result-method}
\alias{exposures<-}
\alias{exposures<-,musica_result,matrix-method}
\title{Retrieve exposures from a musica_result object}
\usage{
exposures(result)

\S4method{exposures}{musica_result}(result)

exposures(result) <- value

\S4method{exposures}{musica_result,matrix}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{value}{A matrix of samples by signature exposures}
}
\value{
A matrix of exposures
}
\description{
The \code{exposure} matrix contains estimated amount of
each signature for each sample. Rows correspond to each signature and
columns correspond to each sample.
}
\examples{
data(res)
exposures(res)
data(res)
exposures(res) <- matrix()
}
