% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{final_comparison}
\alias{final_comparison}
\alias{final_comparison,single_benchmark-method}
\alias{final_comparison<-}
\alias{final_comparison<-,single_benchmark,data.frame-method}
\title{Retrieve the final comparison data.frame from a single_benchmark object}
\usage{
final_comparison(result)

\S4method{final_comparison}{single_benchmark}(result)

final_comparison(result) <- value

\S4method{final_comparison}{single_benchmark,data.frame}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{single_benchmark}} object extracted from
a \code{\linkS4class{full_benchmark}} object}

\item{value}{A data.frame containing the final comparison between predicted
and true signatures, generated by the \link{benchmark_compare_results}
function}
}
\value{
A data.frame object containing the final comparison between predicted
and true signatures
}
\description{
The \code{final_comparison} data.frame contains the comparison
between the predicted signatures and the true signatures after benchmarking
is complete.
}
