% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{final_pred}
\alias{final_pred}
\alias{final_pred,single_benchmark-method}
\alias{final_pred<-}
\alias{final_pred<-,single_benchmark,musica_result-method}
\title{Retrieve the final prediction from a single_benchmark object}
\usage{
final_pred(result)

\S4method{final_pred}{single_benchmark}(result)

final_pred(result) <- value

\S4method{final_pred}{single_benchmark,musica_result}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{single_benchmark}} object extracted from
a \code{\linkS4class{full_benchmark}} object}

\item{value}{A \code{\linkS4class{musica_result}} object containing the
final signature and exposure predictions}
}
\value{
A \code{\linkS4class{musica_result}} object containing the final
predictions
}
\description{
The \code{final_pred} object is a
\code{\linkS4class{musica_result}} object containing the signature and
exposure predictions after benchmarking is complete
}
