% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\docType{class}
\name{full_benchmark-class}
\alias{full_benchmark-class}
\title{Object that contains information for a full benchmarking analysis where
multiple predictions may be benchmarked}
\description{
Object that contains information for a full benchmarking analysis where
multiple predictions may be benchmarked
}
\section{Slots}{

\describe{
\item{\code{ground_truth}}{musica_result}

\item{\code{method_view_summary}}{matrix}

\item{\code{sig_view_summary}}{matrix}

\item{\code{indv_benchmarks}}{list}
}}

