% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_musica.R
\name{get_result_list_entry}
\alias{get_result_list_entry}
\alias{get_result_list_entry,musica-method}
\title{Retrieve result_list from a musica object}
\usage{
get_result_list_entry(object, ...)

\S4method{get_result_list_entry}{musica}(object, entry_name)
}
\arguments{
\item{object}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica} function.}
}
\value{
A list of results
}
\description{
The \code{result_list} contains results from various runs
}
\examples{
data(res)
result_list(res)
}
