% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{get_variants}
\alias{get_variants}
\title{Return variants for musica object}
\usage{
get_variants(musica)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}
}
\value{
Returns variants in musica object
}
\description{
Return variants for musica object
}
\examples{
data(musica)
get_variants(musica)
}
