% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ground_truth}
\alias{ground_truth}
\alias{ground_truth,full_benchmark-method}
\alias{ground_truth<-}
\alias{ground_truth<-,full_benchmark,musica_result-method}
\title{Retrieve ground_truth musica_result object from a full_benchmark object}
\usage{
ground_truth(result)

\S4method{ground_truth}{full_benchmark}(result)

ground_truth(result) <- value

\S4method{ground_truth}{full_benchmark,musica_result}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{full_benchmark}} object generated by
the \link{create_benchmark} function.}

\item{value}{A \code{\linkS4class{musica_result}} object with the ground truth
signatures and exposures}
}
\value{
A \code{\linkS4class{musica_result}} object with the ground truth
signatures and exposures
}
\description{
The \code{ground_truth} musica_result object contains true
signatures and exposures for the benchmarking analysis.
}
