% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{indv_benchmarks}
\alias{indv_benchmarks}
\alias{indv_benchmarks,full_benchmark-method}
\alias{indv_benchmarks<-}
\alias{indv_benchmarks<-,full_benchmark,list-method}
\title{Retrieve indv_benchmarks list from a full_benchmark object}
\usage{
indv_benchmarks(result)

\S4method{indv_benchmarks}{full_benchmark}(result)

indv_benchmarks(result) <- value

\S4method{indv_benchmarks}{full_benchmark,list}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{full_benchmark}} object generated by
the \link{create_benchmark} function.}

\item{value}{A list of \code{\linkS4class{single_benchmark}} objects, each
containing information from a single benchmark run}
}
\value{
A list of \code{\linkS4class{single_benchmark}} objects
}
\description{
The \code{indv_benchmarks} list contains
\code{\linkS4class{single_benchmark}} objects that each contain the benchmarking
information from a single benchmark run.
}
