% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{initial_comparison}
\alias{initial_comparison}
\alias{initial_comparison,single_benchmark-method}
\alias{initial_comparison<-}
\alias{initial_comparison<-,single_benchmark,data.frame-method}
\title{Retrieve the intial comparison data.frame from a single_benchmark object}
\usage{
initial_comparison(result)

\S4method{initial_comparison}{single_benchmark}(result)

initial_comparison(result) <- value

\S4method{initial_comparison}{single_benchmark,data.frame}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{single_benchmark}} object extracted from
a \code{\linkS4class{full_benchmark}} object}

\item{value}{A data.frame containing the initial comparison between predicted
and true signatures, generated by the \link{benchmark_compare_results} function}
}
\value{
A data.frame object containing the initial comparison between predicted
and true signatures
}
\description{
The \code{initial_comparison} data.frame contains the comparison
between the initial predicted signatures and the true signatures.
}
