% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{initial_pred}
\alias{initial_pred}
\alias{initial_pred,single_benchmark-method}
\alias{initial_pred<-}
\alias{initial_pred<-,single_benchmark,musica_result-method}
\title{Retrieve the initial prediction from a single_benchmark object}
\usage{
initial_pred(result)

\S4method{initial_pred}{single_benchmark}(result)

initial_pred(result) <- value

\S4method{initial_pred}{single_benchmark,musica_result}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{single_benchmark}} object extracted from
a \code{\linkS4class{full_benchmark}} object}

\item{value}{A \code{\linkS4class{musica_result}} object containing the
initial signature and exposure predictions}
}
\value{
A \code{\linkS4class{musica_result}} object containing the initial
predictions
}
\description{
The \code{initial_pred} object is a
\code{\linkS4class{musica_result}} object containing the initial signature
and exposure predictions inputted for benchmarking
}
