% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{intermediate_comparison}
\alias{intermediate_comparison}
\alias{intermediate_comparison,single_benchmark-method}
\alias{intermediate_comparison<-}
\alias{intermediate_comparison<-,single_benchmark,data.frame-method}
\title{Retrieve the initial comparison data.frame from a single_benchmark object}
\usage{
intermediate_comparison(result)

\S4method{intermediate_comparison}{single_benchmark}(result)

intermediate_comparison(result) <- value

\S4method{intermediate_comparison}{single_benchmark,data.frame}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{single_benchmark}} object extracted from
a \code{\linkS4class{full_benchmark}} object}

\item{value}{A data.frame containing the intermediate comparison between
predicted and true signatures, generated by the 
\link{benchmark_compare_results} function}
}
\value{
A data.frame object containing the intermediate comparison between
predicted and true signatures
}
\description{
The \code{intermediate_comparison} data.frame contains the comparison
between the predicted signatures and the true signatures after duplicate
signatures have been adjusted but before composite signatures have been
adjusted.
}
