% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{intermediate_pred}
\alias{intermediate_pred}
\alias{intermediate_pred,single_benchmark-method}
\alias{intermediate_pred<-}
\alias{intermediate_pred<-,single_benchmark,musica_result-method}
\title{Retrieve the intermediate prediction from a single_benchmark object}
\usage{
intermediate_pred(result)

\S4method{intermediate_pred}{single_benchmark}(result)

intermediate_pred(result) <- value

\S4method{intermediate_pred}{single_benchmark,musica_result}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{single_benchmark}} object extracted from
a \code{\linkS4class{full_benchmark}} object}

\item{value}{A \code{\linkS4class{musica_result}} object containing the
intermediate signature and exposure predictions}
}
\value{
A \code{\linkS4class{musica_result}} object containing the intermediate
predictions
}
\description{
The \code{intermediate_pred} object is a
\code{\linkS4class{musica_result}} object containing the signature and
exposure predictions after duplicate signatures are adjusted but before
composite sigantures are adjusted
}
