% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{method_id}
\alias{method_id}
\alias{method_id,single_benchmark-method}
\alias{method_id<-}
\alias{method_id<-,single_benchmark,character-method}
\title{Retrieve the method_id from a single_benchmark object}
\usage{
method_id(result)

\S4method{method_id}{single_benchmark}(result)

method_id(result) <- value

\S4method{method_id}{single_benchmark,character}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{single_benchmark}} object extracted from
a \code{\linkS4class{full_benchmark}} object}

\item{value}{A character containing the identifier for the single benchmark
run}
}
\value{
The identifier for the single benchmark run
}
\description{
The \code{method_id} is the identifier for a single
benchmark run.
}
