% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{method_view_summary}
\alias{method_view_summary}
\alias{method_view_summary,full_benchmark-method}
\alias{method_view_summary<-}
\alias{method_view_summary<-,full_benchmark,matrix-method}
\title{Retrieve method_view_summary matrix from a full_benchmark object}
\usage{
method_view_summary(result)

\S4method{method_view_summary}{full_benchmark}(result)

method_view_summary(result) <- value

\S4method{method_view_summary}{full_benchmark,matrix}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{full_benchmark}} object generated by
the \link{create_benchmark} function.}

\item{value}{A matrix containing summary information for all individual
benchmarks run in the analysis}
}
\value{
A matrix containing the summary of all individual benchmarks
}
\description{
The \code{method_view_summary} matrix contains a summary of all
the individual benchmarks completed in the analysis.
}
