% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{musica}
\alias{musica}
\title{musica}
\format{
An object of class \code{musica}
See [create_musica_from_variants()] and [create_musica_from_counts()].
}
\usage{
data(musica)
}
\description{
A musica created for testing that includes SBS variants
}
\keyword{datasets}
