% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmaps using the exposures matrix}
\usage{
plot_heatmap(
  res_annot,
  proportional = FALSE,
  show_column_names = FALSE,
  show_row_names = TRUE,
  scale = TRUE,
  subset_tumor = NULL,
  subset_signatures = NULL,
  annotation = NULL,
  ...
)
}
\arguments{
\item{res_annot}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{proportional}{If \code{TRUE}, then the exposures will be normalized
to between 0 and 1 by dividing by the total number of counts for each sample.
Default \code{FALSE}.}

\item{show_column_names}{Boolean check. If \code{True}, column names are shown. Otherwise, they aren't. 
Default \code{FALSE}}

\item{show_row_names}{Boolean check. If \code{True}, row names are shown. Otherwise, they aren't. 
Default \code{FALSE}}

\item{scale}{Boolean check. If \code{True}, values are scaled by z-score. Otherwise, they aren't.
Default \code{TRUE}}

\item{subset_tumor}{Users can specify certain tumor types on which they want to subset the exposure matrix for 
plotting the heatmap.}

\item{subset_signatures}{Users can specify certain signatures on which they want to subset the exposure matrix 
plotting the heatmap.}

\item{annotation}{Users have the option of plotting the exposure matrix based on their given annotation like
Tumor_Subtypes or age. Error given if the user given annotation doesn't exist in the res_annot annotation object.}

\item{annotations}{Column annotations extracted from the musicatk object}
}
\value{
Generates a heatmap for using the exposure matrix.
}
\description{
The exposures for different signatures can be visualized using a heatmap with this function.
Heatmaps make it easier to visualize the data by representing the magbitude of exposure values
as color in 2-dimensions. The variation in color intensity can help see if the exposures are clustered
or how they vary over space. Exposures can be normalized by providing the \code{proportional} argument.
Column annotations can also be seen by passing the \code{col_annot} argument.
}
\examples{
data(res_annot)
exp <- exposures(res_annot)
ComplexHeatmap::Heatmap(exp)
}
