% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sample_counts}
\alias{plot_sample_counts}
\title{Return sample from musica object}
\usage{
plot_sample_counts(musica, table_name, sample_name)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{table_name}{Name of table used for plotting counts}

\item{sample_name}{Sample name to plot counts}
}
\value{
Generates sample plot {no return}
}
\description{
Return sample from musica object
}
\examples{
musica <- readRDS(system.file("testdata", "musica_sbs96.rds", package = "musicatk"))
plot_sample_counts(musica, "SBS96", get_sample_names(musica)[1])
}
