% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sample_counts}
\alias{plot_sample_counts}
\title{Plot distribution of sample counts}
\usage{
plot_sample_counts(musica, sample_names, table_name = NULL)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{sample_names}{Names of the samples to plot.}

\item{table_name}{Name of table used for plotting counts. If \code{NULL},
then the first table in the \code{\linkS4class{musica}} object will be used.
Default \code{NULL}.}
}
\value{
Generates a ggplot object
}
\description{
Displays the proportion of counts for each mutation type across one
or more samples.
}
\examples{
data(musica_sbs96)
plot_sample_counts(musica_sbs96, sample_names = 
sample_names(musica_sbs96)[1])
}
