% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sample_reconstruction_error}
\alias{plot_sample_reconstruction_error}
\title{Plotting Signature Motif counts/spectra}
\usage{
plot_sample_reconstruction_error(result, sample, plotly = FALSE)
}
\arguments{
\item{result}{S4 Result Object}

\item{sample}{Name or index of sample within result object}

\item{plotly}{add plotly layer for plot interaction}
}
\value{
Generates plot {no return}
}
\description{
Plotting Signature Motif counts/spectra
}
\examples{
result <- readRDS(system.file("testdata", "res.rds", package = "musicatk"))
plot_sample_reconstruction_error(result, "TCGA-ER-A197-06A-32D-A197-08")
}
