% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sample_reconstruction_error}
\alias{plot_sample_reconstruction_error}
\title{Plot reconstruction error for a sample}
\usage{
plot_sample_reconstruction_error(result, sample, plotly = FALSE)
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{sample}{Name of the sample within the
\code{\linkS4class{musica_result}} object.}

\item{plotly}{If \code{TRUE}, the the plot will be made interactive
using \code{\link[plotly]{plotly}}. Default \code{FALSE}.}
}
\value{
Generates a ggplot or plotly object
}
\description{
Displays the observed distribution of counts for each mutation type, 
the distribution of reconstructed counts for each mutation type using 
the inferred mutational signatures, and the difference between the 
two distributions.
}
\examples{
data(res)
plot_sample_reconstruction_error(res, "TCGA-ER-A197-06A-32D-A197-08")
}
