% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_signatures}
\alias{plot_signatures}
\title{Plotting Signature Motif counts/spectra}
\usage{
plot_signatures(
  result,
  legend = TRUE,
  plotly = FALSE,
  color_variable = NULL,
  color_mapping = NULL,
  text_size = 10,
  facet_size = 10,
  show_x_labels = TRUE,
  same_scale = TRUE
)
}
\arguments{
\item{result}{S4 Result Object}

\item{legend}{Whether to include the legend for mutation types in the plot.}

\item{plotly}{add plotly layer for plot interaction}

\item{color_variable}{Annotation column to use for coloring plotted motifs,
provided by counts table from input result's musica object}

\item{color_mapping}{Mapping from color_variable to color names, provided by
counts table from input result's musica object}

\item{text_size}{Size of axis text}

\item{facet_size}{Size of facet text}

\item{show_x_labels}{Toggle plotting of x-axis labels}

\item{same_scale}{If \code{TRUE}, the scale of the probability for each
signature will be the same. If \code{FALSE}, then the scale of the y-axis
will be adjusted for each signature. Default \code{TRUE}.}
}
\value{
Generates plot {no return}
}
\description{
Plotting Signature Motif counts/spectra
}
\examples{
data(res)
plot_signatures(res)
}
