% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{predict_and_benchmark}
\alias{predict_and_benchmark}
\title{Predict and benchmark}
\usage{
predict_and_benchmark(
  musica,
  table_name,
  algorithm = "lda",
  k_min,
  k_max,
  full_benchmark,
  threshold = 0.8,
  adjustment_threshold = 0.9,
  plot = FALSE,
  seed = 1,
  nstart = 10,
  par_cores = 1
)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{table_name}{Name of the table to use for signature discovery. Needs
to be the same name supplied to the table building functions such as
\link{build_standard_table}.}

\item{algorithm}{Method to use for mutational signature discovery. One of 
\code{"lda"} or \code{"nmf"}. Default \code{"lda"}.}

\item{k_min}{Minimum number of singatures to predict}

\item{k_max}{Maximum number of signatures to predict}

\item{full_benchmark}{An object of class \code{\linkS4class{full_benchmark}}
created with the \link{create_benchmark} function or returned from a previous
\code{benchmark} run.}

\item{threshold}{Cosine similarity cutoff for comparing preidcted and true
signatures. Default \code{0.8}.}

\item{adjustment_threshold}{Cosine similarity value of high confidence.
Comparisons that meet this cutoff are assumed to be likely,
while those that fall below the cutoff will be disregarded if the predicted
signature is already captured above the threshold. Default \code{0.9}.}

\item{plot}{If \code{FALSE}, plots will be suppressed. Default \code{TRUE}.}

\item{seed}{Seed to be used for the random number generators in the
signature discovery algorithms. Default \code{1}.}

\item{nstart}{Number of independent random starts used in the mutational
signature algorithms. Default \code{10}.}

\item{par_cores}{Number of parallel cores to use. Only used if
\code{method = "nmf"}. Default \code{1}.}
}
\value{
If \code{make_copy == TRUE}, a new \code{full_benchmark} object is
returned. If \code{make_copy == FALSE}, nothing is returned.
}
\description{
This function will discover signatures from a musica object using a given
algorithm and a range of k values. A new discovery is done for each k value.
As each discovery is completed, the prediction is benchmarked.
}
