% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_prediction.R
\name{predict_exposure}
\alias{predict_exposure}
\title{Prediction of exposures in new samples using pre-existing signatures}
\usage{
predict_exposure(
  musica,
  g,
  table_name,
  signature_res,
  algorithm,
  signatures_to_use = seq_len(ncol(signatures(signature_res))),
  verbose = FALSE
)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{g}{A \linkS4class{BSgenome} object indicating which genome
reference the variants and their coordinates were derived from. Only used
if \code{algorithm = "deconstructSigs"}}

\item{table_name}{Name of table used for posterior prediction.
Must match the table type used to generate the prediction signatures}

\item{signature_res}{Signatures used to predict exposures for the samples
\code{musica} object. Existing signatures need to stored in a
\code{\linkS4class{musica_result}} object.}

\item{algorithm}{Algorithm to use for prediction of exposures. One of
\code{"lda"}, \code{"decompTumor2Sig"}, or
\code{"deconstructSigs"}.}

\item{signatures_to_use}{Which signatures in the \code{signature_res} result
object to use. Default is to use all signatures.}

\item{verbose}{If \code{TRUE}, progress will be printing. Only used if
\code{algorithm = "lda"}. Default \code{FALSE}.}
}
\value{
Returns a A \code{\linkS4class{musica_result}} object containing
signatures given by the \code{signature_res} parameter and exposures
predicted from these signatures.
}
\description{
Exposures for samples will be predicted using an existing set
of signatures stored in a \code{\linkS4class{musica_result}} object. 
Algorithms available for prediction include a modify version of \code{"lda"},
\code{"decompTumor2Sig"}, and \code{"deconstructSigs"}.
}
\examples{
data(musica)
data(cosmic_v2_sigs)
g <- select_genome("19")
build_standard_table(musica, g, "SBS96", overwrite = TRUE)
result <- predict_exposure(musica = musica, table_name = "SBS96",
signature_res = cosmic_v2_sigs, algorithm = "lda")

# Predict using LDA-like algorithm with seed set to 1
set.seed(1)
predict_exposure(musica = musica, table_name = "SBS96",
signature_res = cosmic_v2_sigs, algorithm = "lda")
}
