% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sig_view_summary}
\alias{sig_view_summary}
\alias{sig_view_summary,full_benchmark-method}
\alias{sig_view_summary<-}
\alias{sig_view_summary<-,full_benchmark,matrix-method}
\title{Retrieve sig_view_summary matrix from a full_benchmark object}
\usage{
sig_view_summary(result)

\S4method{sig_view_summary}{full_benchmark}(result)

sig_view_summary(result) <- value

\S4method{sig_view_summary}{full_benchmark,matrix}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{full_benchmark}} object generated by
the \link{create_benchmark} function.}

\item{value}{A matrix containing summary information for all individual
benchmarks run in the analysis, organized by signature}
}
\value{
A matrix containing the summary of all individual benchmarks,
organized by signature
}
\description{
The \code{sig_view_summary} matrix contains a summary of all
the individual benchmarks completed in the analysis, organized by signature
}
