% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{signatures}
\alias{signatures}
\alias{signatures,musica_result-method}
\alias{signatures<-}
\alias{signatures<-,musica_result,matrix-method}
\title{Retrieve signatures from a musica_result object}
\usage{
signatures(result)

\S4method{signatures}{musica_result}(result)

signatures(result) <- value

\S4method{signatures}{musica_result,matrix}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{value}{A matrix of motifs counts by samples}
}
\value{
A matrix of mutational signatures
}
\description{
The \code{signature} matrix contains the probability of
mutation motif in each sample. Rows correspond to each motif and
columns correspond to each signature.
}
\examples{
data(res)
signatures(res)
data(res)
signatures(res) <- matrix()
}
