% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\docType{class}
\name{single_benchmark-class}
\alias{single_benchmark-class}
\title{Object that contains information for a full benchmarking analysis where
multiple predictions may be benchmarked}
\description{
Object that contains information for a full benchmarking analysis where
multiple predictions may be benchmarked
}
\section{Slots}{

\describe{
\item{\code{initial_pred}}{musica_result}

\item{\code{intermediate_pred}}{musica_result}

\item{\code{final_pred}}{musica_result}

\item{\code{initial_comparison}}{data.frame}

\item{\code{intermediate_comparison}}{data.frame}

\item{\code{final_comparison}}{data.frame}

\item{\code{single_summary}}{matrix}

\item{\code{method_id}}{character}

\item{\code{threshold}}{numeric}

\item{\code{adjustment_threshold}}{numeric}

\item{\code{description}}{CharOrNULL}
}}

