% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_class.R
\name{subset_variants_by_samples}
\alias{subset_variants_by_samples}
\title{Return sample from musica_variant object}
\usage{
subset_variants_by_samples(musica, sample_name)
}
\arguments{
\item{musica}{A \code{\linkS4class{musica}} object.}

\item{sample_name}{Sample name to subset by}
}
\value{
Returns sample data.frame subset to a single sample
}
\description{
Return sample from musica_variant object
}
\examples{
musica <- readRDS(system.file("extdata", "musica.rds", package = "musicatk"))
subset_variants_by_samples(musica, "public_LUAD_TCGA-97-7938.vcf")
}
