% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_result_model.R
\name{table_selected}
\alias{table_selected}
\alias{table_selected,result_model-method}
\title{Retrieve table name used for plotting from a result_model object}
\usage{
table_selected(result)

\S4method{table_selected}{result_model}(result)
}
\arguments{
\item{result}{A \code{\linkS4class{result_model}} object generated by
a mutational discovery or prediction tool.}
}
\value{
Table name used for plotting
}
\description{
The table name
}
\examples{
data(res)
model <- get_model(res, "result", "SBS96", "res")
table_selected(model)
}
