% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{threshold}
\alias{threshold}
\alias{threshold,single_benchmark-method}
\alias{threshold<-}
\alias{threshold<-,single_benchmark,numeric-method}
\title{Retrieve the threshold from a single_benchmark object}
\usage{
threshold(result)

\S4method{threshold}{single_benchmark}(result)

threshold(result) <- value

\S4method{threshold}{single_benchmark,numeric}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{single_benchmark}} object extracted from
a \code{\linkS4class{full_benchmark}} object}

\item{value}{The threshold for the single benchmark run}
}
\value{
The threshold for the single benchmark run
}
\description{
The \code{threshold} is the cosine similarity cutoff for
comparisons between predicted and true signatures
}
