% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{umap}
\alias{umap}
\alias{umap,musica_result-method}
\alias{umap<-}
\alias{umap<-,musica_result,matrix-method}
\title{Retrieve umap list from a musica_result object}
\usage{
umap(result)

\S4method{umap}{musica_result}(result)

umap(result) <- value

\S4method{umap}{musica_result,matrix}(result) <- value
}
\arguments{
\item{result}{A \code{\linkS4class{musica_result}} object generated by
a mutational discovery or prediction tool.}

\item{value}{A list of umap dataframes}
}
\value{
A list of umap dataframes
}
\description{
The \code{signature} matrix contains the probability of
mutation motif in each sample. Rows correspond to each motif and
columns correspond to each signature.
}
\examples{
data(res)
umap(res)
data(res)
umap(res) <- matrix()
}
