% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_musica.R
\name{variants}
\alias{variants}
\alias{variants,musica-method}
\alias{variants<-}
\alias{variants<-,musica,data.table-method}
\title{Retrieve variants from a musica or musica_result object}
\usage{
variants(object)

\S4method{variants}{musica}(object)

variants(musica) <- value

\S4method{variants}{musica,data.table}(musica) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica} function or a \code{\linkS4class{musica_result}}
object generated by a mutational discovery or prediction tool.}

\item{musica}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica} function}

\item{value}{A \code{\linkS4class{data.table}} of mutational variants and 
variant-level annotations}
}
\value{
A data.table of variants
}
\description{
The \code{variants} \code{data.table} contains the variants 
and variant-level annotations
}
\examples{
data(res)
variants(res)
data(musica)
variants(musica)
}
