% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_musica.R
\name{variants}
\alias{variants}
\alias{variants,musica-method}
\alias{variants<-}
\alias{variants<-,musica,data.table-method}
\title{Retrieve variants from a musica object}
\usage{
variants(object)

\S4method{variants}{musica}(object)

variants(musica) <- value

\S4method{variants}{musica,data.table}(musica) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica_from_variants} or \link{create_musica_from_counts}
function.}

\item{musica}{A \code{\linkS4class{musica}} object generated by
the \link{create_musica_from_variants} or \link{create_musica_from_counts}
function}

\item{value}{A \code{\linkS4class{data.table}} of mutational variants and
variant-level annotations}
}
\value{
A data.table of variants
}
\description{
The \code{variants} \code{data.table} contains the variants
and variant-level annotations
}
\examples{
data(res)
variants(res)
}
